==> scripts/0_data_preparation.Rmd <==
---
title: 'Data preparation'
author: "Melina Leite"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
classoption: a4paper
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../analysis_matrixBirds") })
output:
  rmdformats::readthedown:
    highlight: kate
    self_contained: true
    thumbnails: false
    lightbox: true
    gallery: false
    fig_heigth: 20
  pdf_document:
    toc: no
header-includes:
- \usepackage{graphicx}
- \usepackage{float}
- \usepackage{setspace}\doublespacing
- \setlength{\parindent}{4em}
#- \usepackage{indentfirst}
- \usepackage{booktabs}
- \usepackage{longtable}
- \usepackage{array}
- \usepackage{multirow}
- \usepackage{colortbl}
- \usepackage{pdflscape}
- \usepackage{tabu}
- \usepackage{threeparttable}
- \usepackage{threeparttablex}
- \usepackage[normalem]{ulem}
- \usepackage{makecell}
- \usepackage{placeins}
- \usepackage[margin=0.8in]{geometry}
documentclass: article
---

\fontsize{11}{11}
\selectfont


```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = T, warning = F, fig.align = "center", message = F,
  error = F, cache = F, fig.pos="H")
#knitr::opts_knit$set(root.dir = "../")

library(knitr)
library(tidyverse); library(GGally)
library(cowplot); theme_set(theme_cowplot())
library(janitor)
library(here)
library(magrittr)
library(betapart)
```

Organization of the datasets for the environmental, species and site/landscape information.
Processing and standardizing the data to create data for each bird assemblage (high and low-quality landscapes, forest specialists and habitat generalists).

The resulting datasets will be used in all subsquent data exploration and analysis.

## Species X sites occurrence matrix

```{r}
data <- read.csv(here("data", "species_site_occ.csv"), check.names=FALSE)
```

Long format
```{r}
cols_to_choose <- names(data[, !names(data) %in% c('landscape', 'site')])
data %<>% tidyr::pivot_longer(cols = all_of(cols_to_choose), 
                              names_to = "sp", 
                              values_to = "occor")
str(data)
## all but the first two - use the names
```

## Sites information

Site information and local forest cover at different buffer radius (400m, 600m, 800m).

```{r}
site <- read.csv(here("data", "sites_information.csv"))
str(site)
```


## Focal landscapes information

landscapes information with land use cover (%) from a buffer at 2 km radius around the centroid of the focal landscape.

```{r}
landscape <- read.csv(here("data", "landscapes_information.csv"))
str(landscape)
```


## Species information and traits

```{r}
sp <- read.csv(here("data","species_information.csv"))
str(sp)
```

## Mergin information

```{r}
dataset <- data %>% 
           left_join(sp[,3:12], "sp") %>% 
           left_join(site [, c(1:3,6:9)],by=c("landscape", "site")) %>%
           left_join(landscape[,c(2,5)], by="landscape")
```

## Creating assemblages datasets 

Cleaning information columns and standardazing continuous variables for modeling.

```{r}
test <- dataset %>% group_by(matrix, habitat_specif) %>%
            select(matrix, habitat_specif,
                   landscape, site, sp, n.visit, occor, 
                   body_size, diet, nest, stratum,
                   frugivory, insectivory, lower_stratum, handwing,
                   forest, 
                   forest_site400, forest_site600, forest_site800) %>%
            mutate(lbody_size = as.numeric(scale(log(body_size))),
                   frugivoryOrig = frugivory,
                   frugivory = as.numeric(scale(frugivory)),
                   insectivoryOrig = insectivory,
                   insectivory = as.numeric(scale(insectivory)),
                   lower_stratumOrig = lower_stratum,
                   lower_stratum = as.numeric(scale(lower_stratum)),
                   handwingOrig = handwing,
                   handwing = as.numeric(scale(handwing)),
                   forest_landOrig = forest,
                   forest_land = as.numeric(scale(forest)),
                   forest_site400Orig = forest_site400,
                   forest_site400 = as.numeric(scale(forest_site400)),
                   forest_site600Orig = forest_site600,
                   forest_site600 = as.numeric(scale(forest_site600)),
                   forest_site800Orig = forest_site800,
                   forest_site800 = as.numeric(scale(forest_site800)),
                   ) %>% as.data.frame()

high.spe <- test %>% filter(matrix=="high_quality",
                               habitat_specif=="specialist") %>%
  select(-matrix,-habitat_specif)
low.spe <- test %>% filter(matrix=="low_quality",
                               habitat_specif=="specialist") %>%
  select(-matrix,-habitat_specif)
high.gen <- test %>% filter(matrix=="high_quality",
                               habitat_specif=="generalist") %>%
  select(-matrix,-habitat_specif)
low.gen <- test %>% filter(matrix=="low_quality",
                               habitat_specif=="generalist") %>%
  select(-matrix,-habitat_specif)
```

Selecting only the species that occur in the landscapes
```{r}
sp.high.spe <- high.spe %>% group_by(sp) %>% summarise(n=sum(occor)) %>% 
  filter(n>0) %>% select(sp)
high.spe <- high.spe %>% filter(sp %in% sp.high.spe$sp)

sp.high.gen <- high.gen %>% group_by(sp) %>% summarise(n=sum(occor)) %>% 
  filter(n>0) %>% select(sp)
high.gen <- high.gen %>% filter(sp %in% sp.high.gen$sp)

sp.low.spe <- low.spe %>% group_by(sp) %>% summarise(n=sum(occor)) %>% 
  filter(n>0) %>% select(sp)
low.spe <- low.spe %>% filter(sp %in% sp.low.spe$sp)

sp.low.gen <- low.gen %>% group_by(sp) %>% summarise(n=sum(occor)) %>% 
  filter(n>0) %>% select(sp)
low.gen <- low.gen %>% filter(sp %in% sp.low.gen$sp)
```

# Summary species diversity

Total richness for forest specialists and habitat generalists in each matrix quality type.
```{r}
riq.high.spe <- length(unique(high.spe$sp))
riq.low.spe <- length(unique(low.spe$sp))
riq.spe <- length(unique(c(unique(low.spe$sp), unique(high.spe$sp))))
shared.spe <- riq.high.spe + riq.low.spe - riq.spe

riq.high.gen <- length(unique(high.gen$sp))
riq.low.gen <- length(unique(low.gen$sp))
riq.gen <- length(unique(c(unique(low.gen$sp), unique(high.gen$sp))))
shared.gen <- riq.high.gen + riq.low.gen - riq.gen  

riq.tot <- riq.gen + riq.spe
riq.tot
```

Richness per landscape and site.
```{r}
# landscape
lcc.spe <- high.spe %>% filter(occor>0) %>% 
  group_by(landscape, sp) %>% 
  summarise(riq = n()) %>%
  group_by(landscape) %>%
  summarise(riq=n()) %>% ungroup() %>%
  summarise(mean = mean(riq), desv.pad = sd(riq),
            median = median(riq),
            max=max(riq), min=min(riq))
lpp.spe <- low.spe %>% filter(occor>0) %>% 
  group_by(landscape,  sp) %>% 
  summarise(riq = n()) %>%
  group_by(landscape) %>%
  summarise(riq=n()) %>% ungroup() %>%
  summarise(mean = mean(riq), desv.pad = sd(riq),
            median = median(riq),
            max=max(riq), min=min(riq))
land.sum <- rbind(lcc.spe, lpp.spe) %>% mutate(matrix = c("High", "low"))

# site
scc.spe <- high.spe %>% filter(occor>0) %>% 
  group_by(landscape, site, forest_site400) %>% 
  summarise(riq = n()) %>% ungroup() %>%
  summarise(mean = mean(riq), desv.pad = sd(riq),
            median = median(riq),
            max=max(riq), min=min(riq))
spp.spe <- low.spe %>% filter(occor>0) %>% 
  group_by(landscape, site, forest_site400) %>% 
  summarise(riq = n()) %>% ungroup() %>%
  summarise(mean = mean(riq), desv.pad = sd(riq),
            median = median(riq),
            max=max(riq), min=min(riq))
site.sum <- rbind(scc.spe, spp.spe) %>% mutate(matrix = c("High", "Low"))

# generalists

# landscape
lcc.gen <- high.gen %>% filter(occor>0) %>% 
  group_by(landscape, sp) %>% 
  summarise(riq = n()) %>%
  group_by(landscape) %>%
  summarise(riq=n()) %>% ungroup() %>%
  summarise(mean = mean(riq), desv.pad = sd(riq),
            median = median(riq),
            max=max(riq), min=min(riq))
lpp.gen <- low.gen %>% filter(occor>0) %>% 
  group_by(landscape,  sp) %>% 
  summarise(riq = n()) %>%
  group_by(landscape) %>%
  summarise(riq=n()) %>% ungroup() %>%
  summarise(mean = mean(riq), desv.pad = sd(riq),
            median = median(riq),
            max=max(riq), min=min(riq))
gland.sum <- rbind(lcc.gen, lpp.gen) %>% mutate(matrix = c("High", "Low"))

# site
scc.gen <- high.gen %>% filter(occor>0) %>% 
  group_by(landscape, site, forest_site400) %>% 
  summarise(riq = n()) %>% ungroup() %>%
  summarise(mean = mean(riq), desv.pad = sd(riq),
            median = median(riq),
            max=max(riq), min=min(riq))
spp.gen <- low.gen %>% filter(occor>0) %>% 
  group_by(landscape, site, forest_site400) %>% 
  summarise(riq = n()) %>% ungroup() %>%
  summarise(mean = mean(riq), desv.pad = sd(riq),
            median = median(riq),
            max=max(riq), min=min(riq))
gsite.sum <- rbind(scc.gen, spp.gen) %>% mutate(matrix = c("High", "Low"))
kable(gsite.sum[,c(6,1:5)])
```

## Beta diversity

Calculating beta diversity among regions of low and high-quality matrix landscapes. 
```{r}
data <- dataset %>% 
  select(matrix,sp,occor) %>% group_by(matrix, sp) %>%
  summarise(occor=sum(occor)) %>%
  pivot_wider(names_from = sp, values_from = occor)
data <- 1*(data[,-1]>0) 

data_spec <- dataset %>% filter(habitat_specif=="specialist") %>%
  select(matrix,sp,occor) %>% group_by(matrix, sp) %>%
  summarise(occor=sum(occor)) %>%
  pivot_wider(names_from = sp, values_from = occor)
data_spec<- 1*(data_spec[,-1]>0) 

data_gen <- dataset %>% filter(habitat_specif=="generalist") %>%
  select(matrix,sp,occor) %>% group_by(matrix, sp) %>%
  summarise(occor=sum(occor)) %>%
  pivot_wider(names_from = sp, values_from = occor)
data_gen<- 1*(data_gen[,-1]>0) 


betas <- data.frame(dataset = c("all_species", "specialists", "generalists"),
                    TotalBeta = c(beta.pair(data)$beta.sor[1],
                             beta.pair(data_spec)$beta.sor[1],
                             beta.pair(data_gen)$beta.sor[1]),
                    Turnover = c(beta.pair(data)$beta.sim[1],
                                     beta.pair(data_spec)$beta.sim[1],
                                     beta.pair(data_gen)$beta.sim[1]),
                    Nestedness = c(beta.pair(data)$beta.sne[1],
                                       beta.pair(data_spec)$beta.sne[1],
                                       beta.pair(data_gen)$beta.sne[1]))
kable(betas, digits=2)


betas2 <- betas %>% select(-TotalBeta) %>% 
  pivot_longer(2:3, names_to = "beta", values_to = "value") %>%
  mutate(dataset = fct_relevel(dataset, "all_species", "specialists", "generalists"))
```


```{r, fig.cap="Beta diversity (all, turnover and nestedness) between assemblages in high- and low-quality matrix landscapes."}
ggplot(betas2, aes(x=dataset, y=value, fill=beta)) +
  geom_col(width=0.6) +
  scale_x_discrete(labels=c("All species", "Specialists", "Generalists"))+
  scale_fill_discrete(name="Beta")+
  xlab("") + ylab("Beta diversity") + ylim(0,1)
```


# Summary traits

## Comparing trait between specialists and generalists

```{r}
bhigh.spe <- high.spe %>% group_by(sp) %>% 
  summarise(body_size = mean(body_size),
            frugi = mean(frugivoryOrig),
            inse  = mean(insectivoryOrig),
            grou = mean(lower_stratumOrig),
            hand = mean(handwingOrig),
            nest = first(nest),
            diet = first(diet),
            stratum =first(stratum)) %>%
  mutate(dataset="high.spe") %>%
  as.data.frame()
blow.spe <- low.spe %>% group_by(sp) %>% 
  summarise(body_size = mean(body_size),
            frugi = mean(frugivoryOrig),
            inse  = mean(insectivoryOrig),
            grou = mean(lower_stratumOrig),
            hand = mean(handwingOrig),
            nest = first(nest),
            diet = first(diet),
            stratum =first(stratum)) %>%
  mutate(dataset="blow.spe") %>%
  as.data.frame()
bhigh.gen <- high.gen %>% group_by(sp) %>% 
  summarise(body_size = mean(body_size),
            frugi = mean(frugivoryOrig),
            inse  = mean(insectivoryOrig),
            grou = mean(lower_stratumOrig),
            hand = mean(handwingOrig),
            nest = first(nest),
            diet = first(diet),
            stratum =first(stratum)) %>%
  mutate(dataset="high.gen") %>%
  as.data.frame()
blow.gen <- low.gen %>% group_by(sp) %>% 
  summarise(body_size = mean(body_size),
            frugi = mean(frugivoryOrig),
            inse  = mean(insectivoryOrig),
            grou = mean(lower_stratumOrig),
            hand = mean(handwingOrig),
            nest = first(nest),
            diet = first(diet),
            stratum =first(stratum)) %>%
  mutate(dataset="low.gen") %>%
  as.data.frame()

bares <- rbind(bhigh.spe, blow.spe, bhigh.gen, blow.gen)
colnames(bares)[2:9] <- c("body mass", "% fruits", "% insects", 
                          "% lower strata","hand-wing index", 
                          "nest type", "main diet", 
                          "foraging strata")
```

```{r}
nbares <- bares %>% separate(dataset, c("matrix", "habitat")) %>% select(-matrix) %>%
  distinct() %>%
  mutate(`body mass` = log(`body mass`)) %>%
  mutate_at(c("body mass", '% fruits', '% insects', '% lower strata',
              "hand-wing index"), scale)

range <- nbares %>% group_by(habitat) %>%
  summarise_if(is.numeric, c(min, max)) %>%
  pivot_longer(2:11, names_to = "trait", values_to = "value") %>%
  separate(trait, c("trait", "es"), "_")
```

```{r,fig.height=7, fig.cap="Boxplots of values for the traits measured as continuous variables for the specialists and generalists. Species from the two regions were grouped."}
nbares %>% pivot_longer(2:6, "trait", "value") %>%
  ggplot(aes(x=habitat, y=value)) + geom_boxplot() +
  geom_text(data=range %>% filter(es =="fn1"), 
            aes(x=habitat, y=value-0.2, label=round(value,2) ))+
    geom_text(data=range %>% filter(es =="fn2"), 
            aes(x=habitat, y=value+0.2, label=round(value,2) ))+
  facet_wrap(~trait) +
  scale_x_discrete(name = "", labels=c("Specialists", "Generalists"))
```



Relationhips among continuous traits for generalists and specialists.

```{r}
ggpairs(sp, columns=c(5,9:12), ggplot2::aes(colour=habitat_specif),
        lower = list(continuous = "smooth"), progress=F) 
```



```{r tab1}
tab1 <- nbares %>% tabyl(`nest type`, habitat) %>% 
  adorn_percentages("col") %>% 
  adorn_pct_formatting(digits = 0) %>% adorn_ns() %>% rename(var=`nest type`)
tab2 <- nbares %>% tabyl(`main diet`, habitat) %>% 
  adorn_percentages("col") %>% 
  adorn_pct_formatting(digits = 0) %>% adorn_ns() %>% rename(var=`main diet`)
tab3 <- nbares %>% tabyl(`foraging strata`, habitat) %>% 
  adorn_percentages("col") %>% 
  adorn_pct_formatting(digits = 0) %>% adorn_ns()  %>% rename(var=`foraging strata`)

rbind(tab1,tab2,tab3) %>% rename(Traits = var, Specialists = spe, Generalists = gen) %>%
  kable()
```




# Saving datasets

```{r}
save(landscape, site, high.spe, high.gen, low.spe, low.gen, 
     file= here("data","datasets.Rdata"))
```
