==> scripts/2_modeling_traits_separate.Rmd <==
---
title: "Models with traits - separate analysis"
author: "Melina Leite"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../") })
output:
  rmdformats::readthedown:
    highlight: kate
    self_contained: true
    thumbnails: false
    lightbox: true
    gallery: false
    fig_heigth: 20
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = T, warning = F, fig.align = "center", message = F,
  error = F, cache = F, fig.pos="H")
knitr::opts_knit$set(root.dir = "../")

library(knitr);library(here)
library(tidyverse); library(cowplot); library(patchwork)
library(lme4); library(bbmle)
theme_set(theme_cowplot())
```


Loading auxiliary function for R2 calculations
```{r}
source(here("scripts","r2_auxiliary_function.R"))
```

Loading data
```{r}
load(here("data","datasets.Rdata"))
# saved models results
load(here("models_outputs","models_traits.Rdata"))
```

```{r}
# objects for all predictions
# local forest cover
f400st.high <- seq(min(high.spe$forest_site400), max(high.spe$forest_site400),
                     length.out=20)
f400Orig.high <- f400st.high * sd(high.spe$forest_site400Orig) +
                                              mean(high.spe$forest_site400Orig)
f400st.low <- seq(min(low.spe$forest_site400), max(low.spe$forest_site400),
                    length.out=20)
f400Orig.low <-f400st.low * sd(low.spe$forest_site400Orig) +
                                              mean(low.spe$forest_site400Orig)

# landscape cover 3 values
lands <- c(20,30,40)
land.high <- (lands- mean(high.spe$forest_landOrig))/sd(high.spe$forest_landOrig)
land.low  <- (lands - mean(low.spe$forest_landOrig))/sd(low.spe$forest_landOrig)
```




# Individual models for each trait

## Body size

```{r, eval=F}
bodyhigh.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*lbody_size + forest_land*lbody_size +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lbody_size|landscape) + (lbody_size|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
bodyhigh.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*lbody_size + forest_land*lbody_size +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lbody_size|landscape) + (lbody_size|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
bodylow.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*lbody_size + forest_land*lbody_size +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lbody_size|landscape) + (lbody_size|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
bodylow.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*lbody_size + forest_land*lbody_size +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lbody_size|landscape) + (lbody_size|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
```

### Predictions

```{r}
# body size 3 values
body <- c(5,100,500)
body.high.spe <- (body - mean(high.spe$body_size))/sd(high.spe$body_size)
body.low.spe  <- (body - mean(low.spe$body_size))/sd(low.spe$body_size)
body.high.gen <- (body - mean(high.gen$body_size))/sd(high.gen$body_size)
body.low.gen  <- (body - mean(low.gen$body_size))/sd(low.gen$body_size)
```

```{r}
# high.spe
pbodyhigh.spe <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             body_size = body)
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       lbody_size = body.high.spe)
pbodyhigh.spe$pred <- predict(bodyhigh.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# high.gen
pbodyhigh.gen <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             body_size = body)
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       lbody_size = body.high.gen)
pbodyhigh.gen$pred <- predict(bodyhigh.gen, newdata = newdata, re.form=NA, 
                         type = "response")
# low.spe
pbodylow.spe <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             body_size = body)
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       lbody_size = body.low.spe)
pbodylow.spe$pred <- predict(bodylow.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# low.gen
pbodylow.gen <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             body_size = body)
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       lbody_size = body.low.gen)
pbodylow.gen$pred <- predict(bodylow.gen, newdata = newdata, re.form=NA, 
                         type = "response")
```

```{r}
pbody <- bind_rows(list(high.spe = pbodyhigh.spe, high.gen = pbodyhigh.gen,
               low.spe = pbodylow.spe, low.gen = pbodylow.gen), .id="group") %>%
  separate(group, c("matrix", "habitat")) %>% 
  mutate(habitat = fct_relevel(habitat, "spe", "gen"))
```

```{r}
labibody <- c(`5`= "5 g", `100`= "100 g", `500`= "500 g", `spe`= "Specialists",
          `gen` = "Generalists")
```

```{r}
figbody <- ggplot(pbody, aes(x=forest_site400, y=pred, col=matrix, 
                  linetype=as.factor(forest_land))) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(body_size), labeller=as_labeller(labibody))+
  scale_linetype_manual(name = "Landscape \n cover (%)", 
                 values = c("dotted",  "dashed","solid")) +
  scale_color_manual(name="Matrix quality", 
                     labels=c("High", "Low"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=14),
        axis.text = element_text(size=12),
        legend.position = "right") +
  ggtitle("Body mass")
figbody
```

```{r, eval=F}
#figure for presentation
jpeg("figures/trait_body_size.jpg", width = 16, height = 10, res=150, units="cm" )
figbody
dev.off()
```

## Nest type

```{r, eval=F}
nesthigh.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*nest + forest_land*nest +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (nest|landscape) + (nest|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
nesthigh.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*nest + forest_land*nest +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (nest|landscape) + (nest|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
nestlow.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*nest + forest_land*nest +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (nest|landscape) + (nest|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
nestlow.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*nest + forest_land*nest +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (nest|landscape) + (nest|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
```

### Predictions

```{r}
# high.spe
pnesthigh.spe <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             nest = unique(high.spe$nest))
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       nest = unique(high.spe$nest))
pnesthigh.spe$pred <- predict(nesthigh.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# high.gen
pnesthigh.gen <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             nest = unique(high.gen$nest))
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       nest = unique(high.gen$nest))
pnesthigh.gen$pred <- predict(nesthigh.gen, newdata = newdata, re.form=NA, 
                         type = "response")
# low.spe
pnestlow.spe <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             nest = unique(low.spe$nest))
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       nest = unique(low.spe$nest))
pnestlow.spe$pred <- predict(nestlow.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# low.gen
pnestlow.gen <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             nest = unique(low.gen$nest))
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       nest = unique(low.gen$nest))
pnestlow.gen$pred <- predict(nestlow.gen, newdata = newdata, re.form=NA, 
                         type = "response")
```

```{r}
pnest <- bind_rows(list(high.spe = pnesthigh.spe, high.gen = pnesthigh.gen,
               low.spe = pnestlow.spe, low.gen = pnestlow.gen), .id="group") %>%
  separate(group, c("matrix", "habitat")) %>% 
  mutate(habitat = fct_relevel(habitat, "spe", "gen"))
```

```{r}
labinest <- c(`cavities`= "Cavities", `closed`= "Closed", 
          `open_semi`= "Open",`spe`= "Specialists",
          `gen` = "Generalists")
```

```{r}
fignest <- ggplot(pnest, aes(x=forest_site400, y=pred, col=matrix, 
                  linetype=as.factor(forest_land))) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(nest), labeller=as_labeller(labinest))+
  scale_linetype_manual(name = "Landscape \n cover (%)", 
                 values = c("dotted",  "dashed","solid")) +
  scale_color_manual(name="Matrix quality", 
                     labels=c("High quality", 
                                              "Low quality"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=14),
        axis.text = element_text(size=12),
        legend.position = "right") +
  ggtitle("Nest type")
fignest
```

```{r, eval=F}
jpeg("figures/trait_nest.jpg", width = 16, height = 10, res=150, units="cm" )
fignest
dev.off()
```


## Diet

```{r, eval=F}
diethigh.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*diet + forest_land*diet +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (diet|landscape) + (diet|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
diethigh.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*diet + forest_land*diet +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (diet|landscape) + (diet|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
dietlow.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*diet + forest_land*diet +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (diet|landscape) + (diet|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
dietlow.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*diet + forest_land*diet +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (diet|landscape) + (diet|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
```

### Predictions

```{r}
# high.spe
pdiethigh.spe <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             diet = unique(high.spe$diet))
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       diet = unique(high.spe$diet))
pdiethigh.spe$pred <- predict(diethigh.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# high.gen
pdiethigh.gen <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             diet = unique(high.gen$diet))
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       diet = unique(high.gen$diet))
pdiethigh.gen$pred <- predict(diethigh.gen, newdata = newdata, re.form=NA, 
                         type = "response")
# low.spe
pdietlow.spe <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             diet = unique(low.spe$diet))
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       diet = unique(low.spe$diet))
pdietlow.spe$pred <- predict(dietlow.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# low.gen
pdietlow.gen <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             diet = unique(low.gen$diet))
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       diet = unique(low.gen$diet))
pdietlow.gen$pred <- predict(dietlow.gen, newdata = newdata, re.form=NA, 
                         type = "response")
```

```{r}
pdiet <- bind_rows(list(high.spe = pdiethigh.spe, high.gen = pdiethigh.gen,
               low.spe = pdietlow.spe, low.gen = pdietlow.gen), .id="group") %>%
  separate(group, c("matrix", "habitat")) %>% 
  mutate(habitat = fct_relevel(habitat, "spe", "gen"))
```

```{r}
labidiet <- c(`insectivorous`= "Insectivorous", `onivorous`= "Onivorous", 
          `frugivorous`= "Frugivorous", `nectarivorous` = "Nectarivorous",
          `granivorous` = "Granivorous",`spe`= "Specialists",
          `gen` = "Generalists")
```

```{r}
figdiet <- ggplot(pdiet, aes(x=forest_site400, y=pred, col=matrix, 
                  linetype=as.factor(forest_land))) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(diet), labeller=as_labeller(labidiet))+
  scale_linetype_manual(name = "Landscape \n cover (%)", 
                 values = c("dotted",  "dashed","solid")) +
  scale_color_manual(name="Matrix quality", 
                     labels=c("High quality", 
                                              "Low quality"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=10, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=10),
        axis.text = element_text(size=10),
        legend.position = "right") +
  ggtitle("diet type")
figdiet
```

```{r, eval=F}
jpeg("figures/trait_diet.jpg", width = 18, height = 12, res=150, units="cm" )
figdiet
dev.off()
```

## % Ground-understory (lower) strata use

```{r, eval=F}
low_strathigh.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*lower_stratum + forest_land*lower_stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lower_stratum|landscape) + (lower_stratum|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
low_strathigh.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*lower_stratum + forest_land*lower_stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lower_stratum|landscape) + (lower_stratum|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
low_stratlow.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*lower_stratum + forest_land*lower_stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lower_stratum|landscape) + (lower_stratum|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
low_stratlow.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*lower_stratum + forest_land*lower_stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lower_stratum|landscape) + (lower_stratum|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
```

### Predictions

```{r}
# low_strat size 3 values
low_strat <- c(0,50,100)
low_strat.high.spe <- (low_strat - mean(high.spe$lower_stratumOrig))/
                                                  sd(high.spe$lower_stratumOrig)
low_strat.low.spe  <- (low_strat - mean(low.spe$lower_stratumOrig))/
                                                   sd(low.spe$lower_stratumOrig)
low_strat.high.gen <- (low_strat - mean(high.gen$lower_stratumOrig))/
                                                  sd(high.gen$lower_stratumOrig)
low_strat.low.gen  <- (low_strat - mean(low.gen$lower_stratumOrig))/
                                                   sd(low.gen$lower_stratumOrig)
```

```{r}
# high.spe
plow_strathigh.spe <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             lower_stratum = low_strat)
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       lower_stratum = low_strat.high.spe)
plow_strathigh.spe$pred <- predict(low_strathigh.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# high.gen
plow_strathigh.gen <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             lower_stratum = low_strat)
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       lower_stratum = low_strat.high.gen)
plow_strathigh.gen$pred <- predict(low_strathigh.gen, newdata = newdata, re.form=NA, 
                         type = "response")
# low.spe
plow_stratlow.spe <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             lower_stratum = low_strat)
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       lower_stratum = low_strat.low.spe)
plow_stratlow.spe$pred <- predict(low_stratlow.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# low.gen
plow_stratlow.gen <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             lower_stratum = low_strat)
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       lower_stratum = low_strat.low.gen)
plow_stratlow.gen$pred <- predict(low_stratlow.gen, newdata = newdata, re.form=NA, 
                         type = "response")
```

```{r}
plow_strat <- bind_rows(list(high.spe = plow_strathigh.spe, high.gen = plow_strathigh.gen,
               low.spe = plow_stratlow.spe, low.gen = plow_stratlow.gen), .id="group") %>%
  separate(group, c("matrix", "habitat")) %>% 
  mutate(habitat = fct_relevel(habitat, "spe", "gen"))
```

```{r}
labi <- c(`0`= "0%", `50`= "50%", `100`= "100%", `spe`= "Specialists",
          `gen` = "Generalists")
```

```{r}
figlow_strat <- ggplot(plow_strat, aes(x=forest_site400, y=pred, col=matrix, 
                  linetype=as.factor(forest_land))) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(lower_stratum), labeller=as_labeller(labi))+
  scale_linetype_manual(name = "Landscape \n cover (%)", 
                 values = c("dotted",  "dashed","solid")) +
  scale_color_manual(name="Matrix quality", 
                     labels=c("High quality", 
                                              "Low quality"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=14),
        axis.text = element_text(size=12),
        legend.position = "right") +
  ggtitle("% Lower stratum use")
figlow_strat
```

```{r, eval=F}
jpeg("figures/trait_low_strat.jpg", width = 16, height = 10, res=150, units="cm" )
figlow_strat
dev.off()
```


## Foraging stratum

```{r, eval=F}
stratumhigh.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*stratum + forest_land*stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (stratum|landscape) + (stratum|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
stratumhigh.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*stratum + forest_land*stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (stratum|landscape) + (stratum|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
stratumlow.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*stratum + forest_land*stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (stratum|landscape) + (stratum|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
stratumlow.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*stratum + forest_land*stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (stratum|landscape) + (stratum|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
```

### Predictions

```{r}
# high.spe
pstratumhigh.spe <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             stratum = unique(high.spe$stratum))
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       stratum = unique(high.spe$stratum))
pstratumhigh.spe$pred <- predict(stratumhigh.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# high.gen
pstratumhigh.gen <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             stratum = unique(high.gen$stratum))
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       stratum = unique(high.gen$stratum))
pstratumhigh.gen$pred <- predict(stratumhigh.gen, newdata = newdata, re.form=NA, 
                         type = "response")
# low.spe
pstratumlow.spe <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             stratum = unique(low.spe$stratum))
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       stratum = unique(low.spe$stratum))
pstratumlow.spe$pred <- predict(stratumlow.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# low.gen
pstratumlow.gen <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             stratum = unique(low.gen$stratum))
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       stratum = unique(low.gen$stratum))
pstratumlow.gen$pred <- predict(stratumlow.gen, newdata = newdata, re.form=NA, 
                         type = "response")
```

```{r}
pstratum <- bind_rows(list(high.spe = pstratumhigh.spe, high.gen = pstratumhigh.gen,
               low.spe = pstratumlow.spe, low.gen = pstratumlow.gen), .id="group") %>%
  separate(group, c("matrix", "habitat")) %>% 
  mutate(habitat = fct_relevel(habitat, "spe", "gen"))
```

```{r}
labi <- c(`ground_under`= "Ground-Understory", `mid_canopy`= "Midstory-Canopy", 
          `all`= "All", `spe`= "Specialists",
          `gen` = "Generalists")
```

```{r}
figstratum <- ggplot(pstratum, aes(x=forest_site400, y=pred, col=matrix, 
                  linetype=as.factor(forest_land))) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(stratum), labeller=as_labeller(labi))+
  scale_linetype_manual(name = "Landscape \n cover (%)", 
                 values = c("dotted",  "dashed","solid")) +
  scale_color_manual(name="Matrix quality", 
                     labels=c("High quality", 
                                              "Low quality"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=14),
        axis.text = element_text(size=12),
        legend.position = "right") +
  ggtitle("stratum type")
figstratum
```

```{r, eval=F}
jpeg("figures/trait_stratum.jpg", width = 16, height = 10, res=150, units="cm" )
figstratum
dev.off()
```




## % Frugivory

```{r, eval=F}
frugihigh.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*frugivory + forest_land*frugivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (frugivory|landscape) + (frugivory|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
frugihigh.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*frugivory + forest_land*frugivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (frugivory|landscape) + (frugivory|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
frugilow.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*frugivory + forest_land*frugivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (frugivory|landscape) + (frugivory|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
frugilow.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*frugivory + forest_land*frugivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (frugivory|landscape) + (frugivory|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
```

### Predictions

```{r}
# frugi size 3 values
frugi <- c(0,50,100)
frugi.high.spe <- (frugi - mean(high.spe$frugivoryOrig))/
                                                  sd(high.spe$frugivoryOrig)
frugi.low.spe  <- (frugi - mean(low.spe$frugivoryOrig))/
                                                   sd(low.spe$frugivoryOrig)
frugi.high.gen <- (frugi - mean(high.gen$frugivoryOrig))/
                                                  sd(high.gen$frugivoryOrig)
frugi.low.gen  <- (frugi - mean(low.gen$frugivoryOrig))/
                                                   sd(low.gen$frugivoryOrig)
```

```{r}
# high.spe
pfrugihigh.spe <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             frugivory = frugi)
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       frugivory = frugi.high.spe)
pfrugihigh.spe$pred <- predict(frugihigh.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# high.gen
pfrugihigh.gen <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             frugivory = frugi)
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       frugivory = frugi.high.gen)
pfrugihigh.gen$pred <- predict(frugihigh.gen, newdata = newdata, re.form=NA, 
                         type = "response")
# low.spe
pfrugilow.spe <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             frugivory = frugi)
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       frugivory = frugi.low.spe)
pfrugilow.spe$pred <- predict(frugilow.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# low.gen
pfrugilow.gen <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             frugivory = frugi)
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       frugivory = frugi.low.gen)
pfrugilow.gen$pred <- predict(frugilow.gen, newdata = newdata, re.form=NA, 
                         type = "response")
```

```{r}
pfrugi <- bind_rows(list(high.spe = pfrugihigh.spe, high.gen = pfrugihigh.gen,
               low.spe = pfrugilow.spe, low.gen = pfrugilow.gen), .id="group") %>%
  separate(group, c("matrix", "habitat")) %>% 
  mutate(habitat = fct_relevel(habitat, "spe", "gen"))
```

```{r}
labi <- c(`0`= "0%", `50`= "50%", `100`= "100%", `spe`= "Specialists",
          `gen` = "Generalists")
```

```{r}
figfrugi <- ggplot(pfrugi, aes(x=forest_site400, y=pred, col=matrix, 
                  linetype=as.factor(forest_land))) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(frugivory), labeller=as_labeller(labi))+
  scale_linetype_manual(name = "Landscape \n cover (%)", 
                 values = c("dotted",  "dashed","solid")) +
  scale_color_manual(name="Matrix quality", 
                     labels=c("High quality", 
                                              "Low quality"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=14),
        axis.text = element_text(size=12),
        legend.position = "right") +
  ggtitle("% Lower stratum use")
figfrugi
```

```{r, eval=F}
jpeg("figures/trait_frugi.jpg", width = 16, height = 10, res=150, units="cm" )
figfrugi
dev.off()
```


## % Insectivory

```{r, eval=F}
insehigh.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*insectivory + forest_land*insectivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (insectivory|landscape) + (insectivory|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
insehigh.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*insectivory + forest_land*insectivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (insectivory|landscape) + (insectivory|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
inselow.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*insectivory + forest_land*insectivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (insectivory|landscape) + (insectivory|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
inselow.gen <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*insectivory + forest_land*insectivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (insectivory|landscape) + (insectivory|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
```

### Predictions

```{r}
# inse size 3 values
inse <- c(0,50,100)
inse.high.spe <- (inse - mean(high.spe$insectivoryOrig))/
                                                  sd(high.spe$insectivoryOrig)
inse.low.spe  <- (inse - mean(low.spe$insectivoryOrig))/
                                                   sd(low.spe$insectivoryOrig)
inse.high.gen <- (inse - mean(high.gen$insectivoryOrig))/
                                                  sd(high.gen$insectivoryOrig)
inse.low.gen  <- (inse - mean(low.gen$insectivoryOrig))/
                                                   sd(low.gen$insectivoryOrig)
```

```{r}
# high.spe
pinsehigh.spe <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             insectivory = inse)
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       insectivory = inse.high.spe)
pinsehigh.spe$pred <- predict(insehigh.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# high.gen
pinsehigh.gen <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands,
                             insectivory = inse)
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       insectivory = inse.high.gen)
pinsehigh.gen$pred <- predict(insehigh.gen, newdata = newdata, re.form=NA, 
                         type = "response")
# low.spe
pinselow.spe <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             insectivory = inse)
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       insectivory = inse.low.spe)
pinselow.spe$pred <- predict(inselow.spe, newdata = newdata, re.form=NA, 
                         type = "response")
# low.gen
pinselow.gen <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land = lands,
                             insectivory = inse)
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       insectivory = inse.low.gen)
pinselow.gen$pred <- predict(inselow.gen, newdata = newdata, re.form=NA, 
                         type = "response")
```

```{r}
pinse <- bind_rows(list(high.spe = pinsehigh.spe, high.gen = pinsehigh.gen,
               low.spe = pinselow.spe, low.gen = pinselow.gen), .id="group") %>%
  separate(group, c("matrix", "habitat")) %>% 
  mutate(habitat = fct_relevel(habitat, "spe", "gen"))
```

```{r}
labi <- c(`0`= "0%", `50`= "50%", `100`= "100%", `spe`= "Specialists",
          `gen` = "Generalists")
```

```{r}
figinse <- ggplot(pinse, aes(x=forest_site400, y=pred, col=matrix, 
                  linetype=as.factor(forest_land))) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(insectivory), labeller=as_labeller(labi))+
  scale_linetype_manual(name = "Landscape \n cover (%)", 
                 values = c("dotted",  "dashed","solid")) +
  scale_color_manual(name="Matrix quality", 
                     labels=c("High quality", 
                                              "Low quality"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=14),
        axis.text = element_text(size=12),
        legend.position = "right") +
  ggtitle("% Lower stratum use")
figinse
```

```{r, eval=F}
jpeg("figures/trait_inse.jpg", width = 16, height = 10, res=150, units="cm" )
figinse
dev.off()
```

## Hand-wing index

```{r, eval=F}
hwhigh.spe <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400*handwing + forest_land*handwing +
         (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                        (handwing|landscape) + (handwing|site),
                      family=binomial, data=high.spe,
                      nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                      optCtrl = list(maxfun = 500000)))

hwlow.spe <- glmer(cbind(occor, n.visit-occor) ~  
                      forest_site400*handwing + forest_land*handwing +
           (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                      (handwing|landscape) + (handwing|site),
                    family=binomial, data=low.spe,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
hwhigh.gen <- glmer(cbind(occor, n.visit-occor) ~  
                      forest_site400*handwing + forest_land*handwing +
                      (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                      (handwing|landscape) + (handwing|site),
                    family=binomial, data=high.gen,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                    optCtrl = list(maxfun = 500000)))

hwlow.gen <- glmer(cbind(occor, n.visit-occor) ~  
                     forest_site400*handwing + forest_land*handwing +
                     (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                     (handwing|landscape) + (handwing|site),
                   family=binomial, data=low.gen,
                   nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                            optCtrl = list(maxfun = 500000)))
```

### Predictions

```{r}
# hand
hand <- c(5,15,35)
hand.high.spe <- (hand- mean(high.spe$handwingOrig, na.rm=T))/
                                sd(high.spe$handwingOrig, na.rm=T)
hand.low.spe <- (hand- mean(low.spe$handwingOrig, na.rm=T))/
                                sd(low.spe$handwingOrig, na.rm=T)
hand.high.gen <- (hand- mean(high.gen$handwingOrig, na.rm=T))/
                                sd(high.spe$handwingOrig, na.rm=T)
hand.low.gen <- (hand- mean(low.gen$handwingOrig, na.rm=T))/
                                sd(low.gen$handwingOrig, na.rm=T)
```

```{r}
# high.spe
phandhigh.spe <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands, 
                             handwing = hand)
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       handwing = hand.high.spe)
phandhigh.spe$pred <- predict(hwhigh.spe, newdata = newdata, re.form=NA, 
                              type = "response")

# low.spe
phandlow.spe <- expand.grid(forest_site400 = f400Orig.low,
                             forest_land =lands, 
                             handwing = hand)
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       handwing = hand.low.spe)
phandlow.spe$pred <- predict(hwlow.spe, newdata = newdata, re.form=NA, 
                             type = "response")


# high.gen
phandhigh.gen <- expand.grid(forest_site400 = f400Orig.high,
                             forest_land = lands, 
                             handwing = hand)
newdata <- expand.grid(forest_site400 = f400st.high,
                       forest_land = land.high,
                       handwing = hand.high.gen)
phandhigh.gen$pred <- predict(hwhigh.gen, newdata = newdata, re.form=NA, 
                              type = "response")

# low.gen
phandlow.gen <- expand.grid(forest_site400 = f400Orig.low,
                            forest_land =lands, 
                            handwing = hand)
newdata <- expand.grid(forest_site400 = f400st.low,
                       forest_land = land.low,
                       handwing = hand.low.gen)
phandlow.gen$pred <- predict(hwlow.gen, newdata = newdata, re.form=NA, 
                             type = "response")
```

```{r}
phand.spe <- bind_rows(list(high.spe = phandhigh.spe, 
                        low.spe = phandlow.spe), .id="group") %>%
  separate(group, c("matrix", "habitat")) 
phand.gen <- bind_rows(list(high.gen = phandhigh.gen, 
                            low.gen = phandlow.gen), .id="group") %>%
  separate(group, c("matrix", "habitat")) 
phand <- bind_rows(phand.spe, phand.gen) %>%
  mutate(habitat = fct_relevel(habitat, "spe", "gen"))
```

```{r}
labihand <- c(`5`= "5", `15`= "15", `35`= "35", `spe`= "Specialists",
              `gen` = "Generalists")
```

```{r}
fighand <- ggplot(phand, aes(x=forest_site400, y=pred, col=matrix,
                                        linetype=as.factor(forest_land))) + 
  geom_line(size=1.5) + 
  facet_grid(habitat~as.factor(handwing), labeller=as_labeller(labihand)) +
   scale_linetype_manual(name = "Landscape \n cover (%)", 
                 values = c("dotted",  "dashed","solid")) +
  scale_color_manual(name="Matrix quality", 
                     labels=c("High", "Low"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2,0.3), limits = c(0,0.35)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=14),
        axis.text = element_text(size=12),
        legend.position = "right") +
  ggtitle("Handwing index")
fighand
```

```{r}
fighand2 <- phand %>% filter(forest_land == 30) %>%
  ggplot(aes(x=forest_site400, y=pred, col=matrix)) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(handwing), labeller=as_labeller(labihand))+
  scale_color_manual(name="  Matrix \n  quality ", 
                     labels=c("High","Low"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2,0.3), limits = c(0,0.35)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=14),
        axis.text = element_text(size=12),
        legend.position = "right",
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        legend.box.background = element_rect(fill="snow2", color="snow2")) +
  ggtitle("Handwing index")
fighand2
```


```{r, eval=F}
jpeg("figures/trait_handwing.jpg", width = 16, height = 10, res=150, units="cm" )
figinse
dev.off()
```

# Comparing trait models

## Specialists high quality matrices

```{r}
AICtab(bodyhigh.spe, diethigh.spe, nesthigh.spe, low_strathigh.spe, hwhigh.spe,
       stratumhigh.spe, frugihigh.spe, insehigh.spe, base=T, weights=T) %>%
  kable()
```

```{r}
r2high.spe <- bind_rows(list(
              body_size = r2.calc(bodyhigh.spe, atrib= "lbody_size"),
              diet = r2.calc(diethigh.spe, atrib= "diet"),
              nest = r2.calc(nesthigh.spe, atrib= "nest"),
              low_strat = r2.calc(low_strathigh.spe, atrib= "lower_stratum"),
              stratum = r2.calc(stratumhigh.spe, atrib= "stratum"),
              frugi = r2.calc(frugihigh.spe, atrib= "frugi"),
              inse_size = r2.calc(insehigh.spe, atrib= "inse"),
              handwing = r2.calc(hwhigh.spe, atrib="handwing")), .id="trait" )
kable(r2high.spe)
```


## Specialists low quality matrices

```{r}
AICtab(bodylow.spe, dietlow.spe, nestlow.spe, low_stratlow.spe, hwlow.spe,
       stratumlow.spe, frugilow.spe, inselow.spe, base=T, weights=T )%>%
  kable()
```

```{r}
r2low.spe <- bind_rows(list(
              body_size = r2.calc(bodylow.spe, atrib= "lbody_size"),
              diet = r2.calc(dietlow.spe, atrib= "diet"),
              nest = r2.calc(nestlow.spe, atrib= "nest"),
              low_strat = r2.calc(low_stratlow.spe, atrib= "lower_stratum"),
              stratum = r2.calc(stratumlow.spe, atrib= "stratum"),
              frugi = r2.calc(frugilow.spe, atrib= "frugi"),
              inse_size = r2.calc(inselow.spe, atrib= "inse"),
              handwing = r2.calc(hwlow.spe, atrib="handwing")), .id="trait" )
kable(r2low.spe)
```


## Generalists high quality matrices

```{r}
AICtab(bodyhigh.gen, diethigh.gen, nesthigh.gen, low_strathigh.gen, hwhigh.gen, 
       stratumhigh.gen, frugihigh.gen, insehigh.gen, base=T, weights=T) %>%
  kable()
```

```{r}
r2high.gen <- bind_rows(list(
              body_size = r2.calc(bodyhigh.gen, atrib= "lbody_size"),
              diet = r2.calc(diethigh.gen, atrib= "diet"),
              nest = r2.calc(nesthigh.gen, atrib= "nest"),
              low_strat = r2.calc(low_strathigh.gen, atrib= "lower_stratum"),
              stratum = r2.calc(stratumhigh.gen, atrib= "stratum"),
              frugi = r2.calc(frugihigh.gen, atrib= "frugi"),
              inse_size = r2.calc(insehigh.gen, atrib= "inse"),
              handwing = r2.calc(hwhigh.gen, atrib="handwing")), .id="trait" )
kable(r2high.gen)
```


## Generalists low quality matrices

```{r}
AICtab(bodylow.gen, dietlow.gen, nestlow.gen, low_stratlow.gen, hwlow.gen,
       stratumlow.gen, frugilow.gen, inselow.gen, base=T, weights=T) %>% 
  kable()
```

```{r}
r2low.gen <- bind_rows(list(
              body_size = r2.calc(bodylow.gen, atrib= "lbody_size"),
              diet = r2.calc(dietlow.gen, atrib= "diet"),
              nest = r2.calc(nestlow.gen, atrib= "nest"),
              low_strat = r2.calc(low_stratlow.gen, atrib= "lower_stratum"),
              stratum = r2.calc(stratumlow.gen, atrib= "stratum"),
              frugi = r2.calc(frugilow.gen, atrib= "frugi"),
              inse_size = r2.calc(inselow.gen, atrib= "inse"),
              handwing = r2.calc(hwlow.gen, atrib="handwing")), .id="trait" )
kable(r2low.gen)
```


# Figure diet + nest type + body size

```{r}
figdiet <- pdiet %>% filter(forest_land == 30) %>%
  ggplot(aes(x=forest_site400, y=pred, col=matrix)) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(diet), labeller=as_labeller(labidiet))+
  scale_color_manual(name="  Matrix \n  quality ", 
                     labels=c("High","Low"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=14),
        axis.text = element_text(size=12),
        legend.position = "right",
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        legend.box.background = element_rect(fill="snow2", color="snow2")) +
  ggtitle("Main diet")

figdiet2 <- figdiet + theme(legend.position = "none")
legenda <- get_legend(figdiet)

figbody <- pbody %>% filter(forest_land == 30) %>%
ggplot(aes(x=forest_site400, y=pred, col=matrix)) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(body_size), labeller=as_labeller(labibody))+
  scale_color_manual(name="Matrix quality", 
                     labels=c("High", "Low"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
         strip.text.y = element_blank(),
        text = element_text(size=14),
        axis.text = element_text(size=12),
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        legend.position = "none") +
  ggtitle("Body mass")

fignest <- pnest %>% filter(forest_land == 30) %>%
ggplot(aes(x=forest_site400, y=pred, col=matrix)) + 
  geom_line(size=1.5) +
  facet_grid(habitat~as.factor(nest), labeller=as_labeller(labinest))+
  scale_color_manual(name="Matrix quality", 
                     labels=c("High", "Low"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=12, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=14),
        axis.text.y = element_blank(),
        axis.text = element_text(size=12),
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        legend.position = "none") +
  ggtitle("Nest type")
```

```{r}
jpeg(here("figures","traits_diet_nest_body.jpg"), width = 20, height = 18, res=150, units="cm" )
(figdiet2+legenda +plot_layout(widths = c(1,0.07)))/(figbody+fignest) 
dev.off()
```


```{r figure 5, echo=F}
figdiet <- pdiet %>% filter(forest_land == 30) %>%
  ggplot(aes(x=forest_site400, y=pred, col=matrix)) + 
  geom_line(size=1) +
  facet_grid(habitat~as.factor(diet), labeller=as_labeller(labidiet))+
  scale_color_manual(name="  Matrix \n  quality ", 
                     labels=c("High","Low"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=10, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=10),
        axis.text = element_text(size=10),
        legend.position = "right",
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        legend.box.background = element_rect(fill="snow2", color="snow2")) +
  ggtitle("Main diet")

figdiet2 <- figdiet + theme(legend.position = "none")
legenda <- get_legend(figdiet)

figbody <- pbody %>% filter(forest_land == 30) %>%
ggplot(aes(x=forest_site400, y=pred, col=matrix)) + 
  geom_line(size=1) +
  facet_grid(habitat~as.factor(body_size), labeller=as_labeller(labibody))+
  scale_color_manual(name="Matrix quality", 
                     labels=c("High", "Low"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="Occurrence probability", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=10, margin = margin(r=2,l=2,t=2,b=2)),
         strip.text.y = element_blank(),
        text = element_text(size=10),
        axis.text = element_text(size=10),
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        legend.position = "none") +
  ggtitle("Body mass")

fignest <- pnest %>% filter(forest_land == 30) %>%
ggplot(aes(x=forest_site400, y=pred, col=matrix)) + 
  geom_line(size=1) +
  facet_grid(habitat~as.factor(nest), labeller=as_labeller(labinest))+
  scale_color_manual(name="Matrix quality", 
                     labels=c("High", "Low"),
                     values=c("darkgreen", "sandybrown")) +
  scale_x_continuous(name="Local forest cover (%)", breaks=c(30,60)) +
  scale_y_continuous(name="", 
                     breaks = c(0,0.1,0.2), limits = c(0,0.2)) +
  theme(strip.text.x = element_text(size=10, margin = margin(r=2,l=2,t=2,b=2)),
        text = element_text(size=10),
        axis.text.y = element_blank(),
        axis.text = element_text(size=10),
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        legend.position = "none") +
  ggtitle("Nest type")

tiff(here("figures","traits_diet_nest_body.tiff"), width = 18, height = 16, res=300, units="cm" )
(figdiet2+legenda +plot_layout(widths = c(1,0.07)))/(figbody+fignest) 
dev.off()

ggsave(here("figures","traits_diet_nest_body.pdf"), device="pdf",
       width=200, height = 180, units="mm")
```

# Trait importance metric

Models withouth trait-enviroment interactions, to calculate trait importance metric.

```{r, eval=F}
bodyhigh.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + lbody_size +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lbody_size|landscape) + (lbody_size|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
bodyhigh.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                         forest_site400 + forest_land + lbody_size +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lbody_size|landscape) + (lbody_size|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
bodylow.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                         forest_site400 + forest_land + lbody_size +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lbody_size|landscape) + (lbody_size|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
bodylow.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                         forest_site400 + forest_land + lbody_size +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lbody_size|landscape) + (lbody_size|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))

nesthigh.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + nest +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (nest|landscape) + (nest|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
nesthigh.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + nest +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (nest|landscape) + (nest|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
nestlow.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                       forest_site400 + forest_land + nest +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (nest|landscape) + (nest|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
nestlow.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                       forest_site400 + forest_land + nest +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (nest|landscape) + (nest|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))

diethigh.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + diet +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (diet|landscape) + (diet|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
diethigh.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + diet +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (diet|landscape) + (diet|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
dietlow.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + diet +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (diet|landscape) + (diet|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
dietlow.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + diet +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (diet|landscape) + (diet|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))

low_strathigh.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + lower_stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lower_stratum|landscape) + (lower_stratum|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
low_strathigh.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                       forest_site400 + forest_land + lower_stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lower_stratum|landscape) + (lower_stratum|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
low_stratlow.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + lower_stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lower_stratum|landscape) + (lower_stratum|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
low_stratlow.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + lower_stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (lower_stratum|landscape) + (lower_stratum|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))

stratumhigh.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (stratum|landscape) + (stratum|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
stratumhigh.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (stratum|landscape) + (stratum|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
stratumlow.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (stratum|landscape) + (stratum|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
stratumlow.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + stratum +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (stratum|landscape) + (stratum|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))

frugihigh.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + frugivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (frugivory|landscape) + (frugivory|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
frugihigh.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + frugivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (frugivory|landscape) + (frugivory|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
frugilow.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + frugivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (frugivory|landscape) + (frugivory|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
frugilow.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + frugivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (frugivory|landscape) + (frugivory|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))

insehigh.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + insectivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (insectivory|landscape) + (insectivory|site),
                       family=binomial, data=high.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
insehigh.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + insectivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (insectivory|landscape) + (insectivory|site),
                       family=binomial, data=high.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
inselow.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + insectivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (insectivory|landscape) + (insectivory|site),
                       family=binomial, data=low.spe,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))
inselow.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                        forest_site400 + forest_land + insectivory +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                         (insectivory|landscape) + (insectivory|site),
                       family=binomial, data=low.gen,
                       nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                           optCtrl = list(maxfun = 500000)))

hwhigh.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                      forest_site400 + forest_land +handwing +
                      (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                      (handwing|landscape) + (handwing|site),
                    family=binomial, data=high.spe,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                             optCtrl = list(maxfun = 500000)))

hwlow.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                     forest_site400 + forest_land + handwing +
             (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                     (handwing|landscape) + (handwing|site),
                   family=binomial, data=low.spe,
                   nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                            optCtrl = list(maxfun = 500000)))

hwhigh.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                       forest_site400 + forest_land +handwing +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                       (handwing|landscape) + (handwing|site),
                     family=binomial, data=high.gen,
                     nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                            optCtrl = list(maxfun = 500000)))

hwlow.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                      forest_site400 + forest_land + handwing +
            (forest_site400 + forest_land|sp) + (1|landscape:sp) + (1|site:sp) + 
                      (handwing|landscape) + (handwing|site),
                    family=binomial, data=low.gen,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                             optCtrl = list(maxfun = 500000)))
```



# Saving models results

```{r, eval=F}
save(bodyhigh.spe, bodyhigh.gen, bodylow.spe, bodylow.gen,
     diethigh.spe, diethigh.gen, dietlow.spe, dietlow.gen,
     nesthigh.spe, nesthigh.gen, nestlow.spe, nestlow.gen,
     low_strathigh.spe, low_strathigh.gen, low_stratlow.spe, low_stratlow.gen,
     stratumhigh.spe, stratumhigh.gen, stratumlow.spe, stratumlow.gen,
     frugihigh.spe, frugihigh.gen, frugilow.spe,frugilow.gen,
     insehigh.spe, insehigh.gen, inselow.spe, inselow.gen,
     hwhigh.spe, hwhigh.gen, hwlow.spe, hwlow.gen,
     # no interaction trait*env
     bodyhigh.spe2, bodyhigh.gen2, bodylow.spe2, bodylow.gen2,
     diethigh.spe2, diethigh.gen2, dietlow.spe2, dietlow.gen2,
     nesthigh.spe2, nesthigh.gen2, nestlow.spe2, nestlow.gen2,
     low_strathigh.spe2, low_strathigh.gen2, low_stratlow.spe2, low_stratlow.gen2,
     stratumhigh.spe2, stratumhigh.gen2, stratumlow.spe2, stratumlow.gen2,
     frugihigh.spe2, frugihigh.gen2, frugilow.spe2,frugilow.gen2,
     insehigh.spe2, insehigh.gen2, inselow.spe2, inselow.gen2,
     hwhigh.spe2, hwhigh.gen2, hwlow.spe2, hwlow.gen2,
     file= here("models_outputs","models_traits.Rdata"))
```
