==> scripts/3_modeling_traits_combined.Rmd <==
---
title: "Final models traits - forest cover at local and landscape scale"
author: "Melina Leite"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../") })
output:
  rmdformats::readthedown:
    highlight: kate
    self_contained: true
    thumbnails: false
    lightbox: true
    gallery: false
    fig_heigth: 20
  pdf_document:
    highlight: tango
    toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = T, warning = F, fig.align = "center", message = F,
  error = F, cache = F, fig.pos="H")
knitr::opts_knit$set(root.dir = "../")

library(knitr); library(here)
library(tidyverse); library(cowplot); library(patchwork); library(ggrepel)
library(lme4); library(bbmle); library(broom.mixed)
theme_set(theme_cowplot())
```

Loading auxiliary function for R2 calculations
```{r}
source(here("scripts","r2_auxiliary_function.R"))
```

Loading data
```{r}
load(here("data","datasets.Rdata"))
load(here("models_outputs","models_combined.Rdata")) # saved models results this script
load(here("models_outputs","models_traits.Rdata")) # saved models results separate
```


# Combined traits models

```{r, eval=F}
mhigh.spe <- glmer(cbind(occor, n.visit-occor) ~  
                    forest_site400*lbody_size + 
                    forest_site400*nest +
                    forest_site400*diet + 
                    forest_site400*lower_stratum +
                    forest_land*lbody_size + 
                    forest_land*nest + 
                    forest_land*diet +
                    forest_land*lower_stratum + 
                    (forest_site400 + forest_land|sp) + 
                    (1|landscape:sp) + (1|site:sp) + 
                    (lbody_size + nest + diet + lower_stratum|landscape) +
                    (lbody_size + nest + diet + lower_stratum|site),
                    family=binomial, data = high.spe,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                    optCtrl = list(maxfun = 500000)))


mlow.spe <- glmer(cbind(occor, n.visit-occor) ~  
                    forest_site400*lbody_size + 
                    forest_site400*nest +
                    forest_site400*diet + 
                    forest_site400*lower_stratum +
                    forest_land*lbody_size + 
                    forest_land*nest + 
                    forest_land*diet +
                    forest_land*lower_stratum + 
                    (forest_site400 + forest_land|sp) + 
                    (1|landscape:sp) + (1|site:sp) + 
                    (lbody_size + nest + diet + lower_stratum|landscape) +
                    (lbody_size + nest + diet + lower_stratum|site),
                    family=binomial, data=low.spe,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                    optCtrl = list(maxfun = 500000)))


mhigh.gen <- glmer(cbind(occor, n.visit-occor) ~  
                    forest_site400*lbody_size + 
                    forest_site400*nest +
                    forest_site400*diet + 
                    forest_site400*lower_stratum +
                    forest_land*lbody_size + 
                    forest_land*nest + 
                    forest_land*diet +
                    forest_land*lower_stratum + 
                    (forest_site400 + forest_land|sp) + 
                    (1|landscape:sp) + (1|site:sp) + 
                    (lbody_size + nest + diet + lower_stratum|landscape) +
                    (lbody_size + nest + diet + lower_stratum|site),
                    family=binomial, data=high.gen,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                    optCtrl = list(maxfun = 500000)))


mlow.gen <- glmer(cbind(occor, n.visit-occor) ~  
                    forest_site400*lbody_size + 
                    forest_site400*nest +
                    forest_site400*diet + 
                    forest_site400*lower_stratum +
                    forest_land*lbody_size + 
                    forest_land*nest + 
                    forest_land*diet +
                    forest_land*lower_stratum + 
                    (forest_site400 + forest_land|sp) + 
                    (1|landscape:sp) + (1|site:sp) + 
                    (lbody_size + nest + diet + lower_stratum|landscape) +
                    (lbody_size + nest + diet + lower_stratum|site),
                    family=binomial, data=low.gen,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                    optCtrl = list(maxfun = 500000)))
```

# Local and landscape forest cover effects

Comparing effect sizes and 95% IC for fixed effects forest cover a local and landscape scale for each assemblage.

```{r, echo=F}
fhigh.spe <- tidy(mhigh.spe, effects = "fixed") %>% select(-effect)
fhigh.gen <- tidy(mhigh.gen, effects = "fixed") %>% select(-effect)
flow.spe  <- tidy(mlow.spe, effects = "fixed") %>% select(-effect)
flow.gen  <- tidy(mlow.gen, effects = "fixed") %>% select(-effect)
```

```{r, echo=F}
fc <- bind_rows(list(High.Specialists=fhigh.spe, 
                     Low.Specialists=flow.spe, 
                     High.Generalists=fhigh.gen,
                     Low.Generalists =flow.gen), .id= "model") %>%
  separate(model, c("matrix", "habitat"), remove = F) %>%
  filter(term %in% c("forest_land", "forest_site400")) %>%
  arrange(habitat, matrix,term) %>%
  mutate(iclow = estimate - 1.96*std.error,
         icup = estimate + 1.96*std.error, 
         habitat = fct_relevel(habitat, "Specialists", "Generalists"),
         term = fct_recode(term, Landscape = "forest_land", Local = "forest_site400")) %>%
  mutate(term = fct_relevel(term, "Landscape", "Local"),
         matrix = fct_relevel(matrix, "Low", "High"))

fig.forest <- fc %>%
  ggplot(aes(x=estimate, y=term, col=term)) + geom_point(size=3) + 
  geom_errorbarh(aes(xmin=iclow, xmax=icup), height=0, size=1.5) +
  facet_grid(cols= vars(matrix), rows=vars(habitat)) +
  geom_vline(xintercept = 0, linetype="dashed", col="grey") +
  scale_color_manual(name="  Forest \n  cover ",
                     values=c( "yellow4", "turquoise4")) +
 theme(text = element_text(size=12),
        axis.text = element_text(size=12),
        strip.text.x = element_text(size=12, margin=margin(t=2,b=3)),
        strip.text.y = element_text(size=12, margin=margin(r=2,l=3)),
       #legend.position = c(0.89, 0.2),
       plot.title = element_text(hjust = 0.5),
       legend.position = "none",
       panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA")
       ) +
  xlab("Coefficient (slope)") + ylab("Forest cover")+
  ggtitle("Matrix quality")
fig.forest
```

```{r, echo=F, results="hide"}
jpeg(here("figures","forest_cover_coefs.jpg"), width=16, height = 10, units="cm", res=1000)
fig.forest
dev.off()
```

```{r figure 3, echo=F, results="hide"}
# figure for publication
fig.forest2 <- fc %>%
  ggplot(aes(x=estimate, y=term, col=term)) + 
  geom_vline(xintercept = 0, linetype="dashed", col="grey", size=0.3) +
  geom_point(size=1) + 
  geom_errorbarh(aes(xmin=iclow, xmax=icup), height=0, size=0.5) +
  facet_grid(cols= vars(matrix), rows=vars(habitat)) +
  scale_color_manual(name="  Forest \n  cover ",
                     values=c( "yellow4", "turquoise4")) +
 theme(text = element_text(size=6),
        axis.text = element_text(size=6),
        strip.text.x = element_text(size=6, margin=margin(t=1,b=2)),
        strip.text.y = element_text(size=6, margin=margin(r=1,l=2)),
       axis.line = element_line(size=0.3),
       axis.ticks = element_line(size=0.3),
       plot.title = element_text(hjust = 0.5),
       legend.position = "none",
       panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA")
       ) +
  xlab("Coefficient (slope)") + ylab("Forest cover")+
  ggtitle("Matrix quality")
ggsave(here("figures","forest_cover_coefs.pdf"), device="pdf", width=80,
       height = 60, units="mm")
tiff(here("figures","forest_cover_coefs.tiff"), width=8, height = 6, units="cm", res=600)
fig.forest2
dev.off()
```


# Trait importance metric

The importance of the trait in explaining habitat loss filtering was calculated using the proportional decrease in the variance of the species random slope of forest cover ($d_{spp[i]$) when you include the trait-environment interaction term ($\beta_{12}$) in the model (Jamil et al. 2013):

$$C_\beta = 1 = \frac{d_{spp[i]}(res)}{d_{spp[i]}(total)}$$




where $d_{spp[i]}(res)$ is the variance of the random slope from model with trait main effects ($\beta_{2}$) and the trait-environment interaction ($\beta_{12}$), and $d_{spp[i]}(total)$ is the same variance term for the model without trait-environment interaction but with the main effects of traits.This proportion explains how much of the effect of habitat loss on each species can be explained by their trait values.

**OBS:** Models without trait*enviroment interaction for the separate traits were previoulsy analyzed in the Rmd scrit `3_modeling_traits_separate.Rmd`.

Models without trait*enviroment interactions for the combined traits:

```{r, eval=F}
mhigh.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                    forest_site400 + forest_land +
                    lbody_size + 
                    nest +
                    diet + 
                    lower_stratum +
                    (forest_site400 + forest_land|sp) + 
                    (1|landscape:sp) + (1|site:sp) + 
                    (lbody_size + nest + diet + lower_stratum|landscape) +
                    (lbody_size + nest + diet + lower_stratum|site),
                    family=binomial, data=high.spe,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                    optCtrl = list(maxfun = 500000)))


mlow.spe2 <- glmer(cbind(occor, n.visit-occor) ~  
                    forest_site400 + forest_land +
                    lbody_size + 
                    nest +
                    diet + 
                    lower_stratum +
                    (forest_site400 + forest_land|sp) + 
                    (1|landscape:sp) + (1|site:sp) + 
                    (lbody_size + nest + diet + lower_stratum|landscape) +
                    (lbody_size + nest + diet + lower_stratum|site),
                    family=binomial, data=low.spe,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                    optCtrl = list(maxfun = 500000)))


mhigh.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                    forest_site400 + forest_land +
                    lbody_size + 
                    nest +
                    diet + 
                    lower_stratum +
                    (forest_site400 + forest_land|sp) + 
                    (1|landscape:sp) + (1|site:sp) + 
                    (lbody_size + nest + diet + lower_stratum|landscape) +
                    (lbody_size + nest + diet + lower_stratum|site),
                    family=binomial, data=high.gen,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                    optCtrl = list(maxfun = 500000)))


mlow.gen2 <- glmer(cbind(occor, n.visit-occor) ~  
                    forest_site400 + forest_land +
                    lbody_size + 
                    nest +
                    diet + 
                    lower_stratum +
                    (forest_site400 + forest_land|sp) + 
                    (1|landscape:sp) + (1|site:sp) + 
                    (lbody_size + nest + diet + lower_stratum|landscape) +
                    (lbody_size + nest + diet + lower_stratum|site),
                    family=binomial, data=low.gen,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                    optCtrl = list(maxfun = 500000)))

save(mhigh.gen, mhigh.spe, mlow.gen, mlow.spe,
     mhigh.gen2, mhigh.spe2, mlow.gen2, mlow.spe2,
  file="models_outputs/models_combined.Rdata")
```

Function to calculate trait importance:
```{r}
cbeta <- function(m.inter, m.no.inter){
    cbeta <- 1-((VarCorr(m.inter)$sp[2,2] + VarCorr(m.inter)$sp[3,3])
               /(VarCorr(m.no.inter)$sp[2,2] + VarCorr(m.no.inter)$sp[3,3]))
  return(round(100*cbeta))
}
```

```{r}
comb <- c(cbeta(mhigh.spe, mhigh.spe2),
          cbeta(mlow.spe,  mlow.spe2),
          cbeta(mhigh.gen, mhigh.gen2),
          cbeta(mlow.gen,  mlow.gen2))

bd   <- c(cbeta(bodyhigh.spe, bodyhigh.spe2),
          cbeta(bodylow.spe,  bodylow.spe2),
          cbeta(bodyhigh.gen, bodyhigh.gen2),
          cbeta(bodylow.gen,  bodylow.gen2))

ni   <- c(cbeta(nesthigh.spe, nesthigh.spe2),
          cbeta(nestlow.spe,  nestlow.spe2),
          cbeta(nesthigh.gen, nesthigh.gen2),
          cbeta(nestlow.gen,  nestlow.gen2))

di   <- c(cbeta(diethigh.spe, diethigh.spe2),
          cbeta(dietlow.spe,  dietlow.spe2),
          cbeta(diethigh.gen, diethigh.gen2),
          cbeta(dietlow.gen,  dietlow.gen2))

ls   <- c(cbeta(low_strathigh.spe, low_strathigh.spe2),
          cbeta(low_stratlow.spe,  low_stratlow.spe2),
          cbeta(low_strathigh.gen, low_strathigh.gen2),
          cbeta(low_stratlow.gen,  low_stratlow.gen2))

st   <- c(cbeta(stratumhigh.spe, stratumhigh.spe2),
          cbeta(stratumlow.spe,  stratumlow.spe2),
          cbeta(stratumhigh.gen, stratumhigh.gen2),
          cbeta(stratumlow.gen,  stratumlow.gen2))

fr   <- c(cbeta(frugihigh.spe, frugihigh.spe2),
          cbeta(frugilow.spe,  frugilow.spe2),
          cbeta(frugihigh.gen, frugihigh.gen2),
          cbeta(frugilow.gen,  frugilow.gen2))

is   <- c(cbeta(insehigh.spe, insehigh.spe2),
          cbeta(inselow.spe,  inselow.spe2),
          cbeta(insehigh.gen, insehigh.gen2),
          cbeta(inselow.gen,  inselow.gen2))

hw <- c(cbeta(hwhigh.spe, hwhigh.spe2),
          cbeta(hwlow.spe,  hwlow.spe2),
          cbeta(hwhigh.gen, hwhigh.gen2),
          cbeta(hwlow.gen,  hwlow.gen2))
```

```{r}
tra <- data.frame(matrix = c("High quality", "Low quality", 
                             "High quality", "Low quality"),
                  habitat = c("Specialists", "Specialists",
                              "Generalists","Generalists"),
                  comb, ni, di, st, is, fr, ls, bd, hw)
colnames(tra)[3:11] <- c("Combined","Nest type", "Main diet", 
        "Foraging stratum", "% insectivory", "% frugivory", "% lower strata",
                             "Body mass", "handwing")
kable(tra)
```

```{r}
figtrait.env <- tra %>% gather("trait", "value", 3:11) %>%
  mutate(trait = fct_relevel(trait, "Combined","Nest type", "Main diet", 
                             "Foraging stratum", "% insectivory", "% frugivory", 
                             "% lower strata",
                             "Body mass"),
         habitat = fct_relevel(habitat, "Specialists", "Generalists")) %>% 
  arrange(trait) %>%
  mutate(xis=rep(1:9, each=4)) %>%
  ggplot(aes(x=xis, y=value, col=matrix)) + 
    geom_line(size=1) + 
    geom_point(size=3) +
    facet_grid(~habitat) +
    scale_x_continuous(breaks=1:9, labels= c("Combined","Nest type", "Main diet", 
                                           "Foraging stratum", "% insectivory", 
                                           "% frugivory", "% lower strata",
                                           "Body mass", "Hand-wing")) +
  xlab("") +
  scale_color_manual(name="Matrix \n quality",
                     labels = c("High", "Low"),
                     values = c("darkgreen", "sandybrown")) +
  ylab("Trait importance for \n habitat loss filtering (%)")+
  theme(axis.text.x = element_text(angle=45, hjust=1),
        text = element_text(size=12),
        axis.text = element_text(size=12),
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        legend.box.background = element_rect(fill="snow2", color="snow2"),
        strip.text = element_text(size=12, margin=margin(t=2,b=3)),
        legend.justification = c(1, 0), legend.position = c(1, 0.65)) 
figtrait.env
```

```{r}
jpeg(here("figures","fig_traitXenv.jpg"), width=20, height = 12, units="cm", res=150)
figtrait.env
dev.off()
```

```{r figure 4, echo=F}
# figure for publication
figtrait.env <- tra %>% gather("trait", "value", 3:11) %>%
  mutate(trait = fct_relevel(trait, "Combined","Nest type", "Main diet", 
                             "Foraging stratum", "% insectivory", "% frugivory", 
                             "% lower strata",
                             "Body mass"),
         habitat = fct_relevel(habitat, "Specialists", "Generalists")) %>% 
  arrange(trait) %>%
  mutate(xis=rep(1:9, each=4)) %>%
  ggplot(aes(x=xis, y=value, col=matrix)) + 
    geom_line(size=1) + 
    geom_point(size=2) +
    facet_grid(~habitat) +
    scale_x_continuous(breaks=1:9, labels= c("Combined","Nest type", "Main diet", 
                                           "Foraging stratum", "% insectivory", 
                                           "% frugivory", "% lower strata",
                                           "Body mass", "Hand-wing")) +
  xlab("") + ylim(0,30)+
  scale_color_manual(name="Matrix \n quality",
                     labels = c("High", "Low"),
                     values = c("darkgreen", "sandybrown")) +
  ylab("Trait importance for \n habitat loss filtering (%)")+
  theme(axis.text.x = element_text(angle=45, hjust=1),
        text = element_text(size=10),
        axis.text = element_text(size=10),
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        legend.box.background = element_rect(fill="snow2", color="snow2"),
        strip.text = element_text(size=10, margin=margin(t=2,b=3)),
        legend.justification = c(1, 0), legend.position = c(1, 0.55)) 
ggsave(here("figures","fig_traitXenv.pdf"), device="pdf", width=180,
       height = 100, units="mm")
tiff(here("figures","fig_traitXenv.tiff"), width=18, height = 10, units="cm", res=300)
figtrait.env
dev.off()
```


# R^2^ models

```{r}
traits = c("lbody_size", "nest", "diet", "lower_stratum")
vals <- bind_rows(list(High_Specialists = r2.calc(mhigh.spe, atrib=traits)*100,
      Low_Specialists=r2.calc(mlow.spe, atrib=traits)*100,
      High_Generalists=r2.calc(mhigh.gen, atrib=traits)*100,
     Low_Generalists=r2.calc(mlow.gen, atrib=traits)*100), .id="assemblage")

colnames(vals) <- c("assemblage", "condicional", "trait*env", "env|sp", "lands:sp", "site:sp", "trait|lands", "trait|site")

vals2 <-  vals %>%
  separate(assemblage, c("matrix", "habitat")) %>%
  select(-condicional) %>%
  pivot_longer(3:8, names_to = "term", values_to = "R2") %>%
  mutate(term = fct_relevel(term,"trait|site", "trait|lands", "site:sp", 
                                  "lands:sp", "trait*env","env|sp"),
         habitat = fct_relevel(habitat, "Specialists", "Generalists"),
         matrix = fct_relevel(matrix, "Low", "High")) %>%
  arrange(habitat,matrix,desc(term)) %>%
  group_by(habitat, matrix) %>%
  mutate(yis = cumsum(R2)-R2/2, 
         label = as.character(round(R2))) %>% ungroup() %>%
  mutate(xis = rep(rep(c(1.31,2.31), each=6), 2))

fig.part <- ggplot(vals2, aes(x=matrix, y=R2, fill=term)) + geom_col(width=0.7) +
  scale_fill_manual(name="Model terms" ,
                    values = c("lightgrey","darkgrey","aquamarine1",
                               "aquamarine3", "slateblue1", "slateblue4")) +
  facet_wrap(~habitat)+
  scale_y_continuous(name= expression(~ R^2), breaks=seq(0,70,10),
                     limits=c(0,75)) +
  scale_x_discrete(name="Matrix quality") +
  theme(text = element_text(size=12),
        axis.text = element_text(size=12),
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        strip.text = element_text(size=12, margin=margin(t=2,b=3))) +
geom_text_repel(data=vals2, aes(label=label, y=yis, x=xis), 
                segment.size = 0.15, size=3,
                segment.alpha = 0.7,
                direction="y",
                show.legend=F,   nudge_x = 0.15, hjust=0)

fig.part
```


```{r}
jpeg(here("figures", "fig_R2.jpg"), width=20, height = 12, units="cm", res=150)
fig.part
dev.off()
```

```{r figure 2, echo=F}
fig.part <- ggplot(vals2, aes(x=matrix, y=R2, fill=term)) + geom_col(width=0.7) +
  scale_fill_manual(name="Model terms" ,
                    values = c("lightgrey","darkgrey","aquamarine1",
                               "aquamarine3", "slateblue1", "slateblue4")) +
  facet_wrap(~habitat)+
  scale_y_continuous(name= expression(~ R^2), breaks=seq(0,70,10),
                     limits=c(0,75)) +
  scale_x_discrete(name="Matrix quality") +
  theme(text = element_text(size=12),
        axis.text = element_text(size=12),
        panel.background = element_rect(fill="#FAFAFA", color="#FAFAFA"),
        strip.text = element_text(size=12, margin=margin(t=2,b=3))) +
geom_text_repel(data=vals2, aes(label=label, y=yis, x=xis), 
                segment.size = 0.15, size=3,
                segment.alpha = 0.7,
                direction="y",
                show.legend=F,   nudge_x = 0.15, hjust=0)
ggsave(here("figures","fig_R2.pdf"), device="pdf", width=180,
       height = 100, units="mm")
tiff(here("figures","fig_R2.tiff"), width=18, height = 10, units="cm", res=300)
fig.part
dev.off()
```
