==> scripts/4_models_diagnostic_appendix3.Rmd <==
---
title: "Appendix S3: Additional models results"
author: "Melina Leite"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
bibliography: ../references/refs_appendix.bib
csl: ../references/references-style.csl
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../") })
output:
  pdf_document:
    highlight: zenburn
    toc: no
  rmdformats::readthedown:
    highlight: kate
    self_contained: true
    thumbnails: false
    lightbox: true
    gallery: false
    fig_heigth: 20
documentclass: article
classoption: a4paper
header-includes:
    - \renewcommand{\figurename}{Figure S3.}
    - \makeatletter
    - \def\fnum@figure{\figurename\thefigure}
    - \makeatother
    - \renewcommand{\tablename}{Table S3.}
    - \makeatletter
    - \def\fnum@table{\tablename\thetable}
    - \makeatother
    - \usepackage{graphicx}
    - \usepackage{float}
    - \usepackage{setspace}\doublespacing
    - \setlength{\parindent}{4em}
    - \usepackage{indentfirst}
    - \usepackage{booktabs}
    - \usepackage{longtable}
    - \usepackage{array}
    - \usepackage{multirow}
    #- \usepackage[table]{xcolor}
    - \usepackage{colortbl}
    - \usepackage{pdflscape}
    - \usepackage{tabu}
    - \usepackage{threeparttable}
    - \usepackage{threeparttablex}
    - \usepackage[normalem]{ulem}
    - \usepackage{makecell}
    - \usepackage{placeins}
    - \usepackage[margin=0.8in]{geometry}
---

\fontsize{11}{11}
\selectfont


```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = F, warning = F, fig.align = "center", message = F,
  error = F, cache = F, fig.pos="H")
knitr::opts_knit$set(root.dir = "../")

library(knitr)
library(here)
library(corrplot)
library(tidyverse); library(cowplot); library(patchwork); #theme_set(theme_cowplot)
library(lme4); library(bbmle); library(broom.mixed)
options(kableExtra.latex.load_packages = FALSE)
library(kableExtra)
library(DHARMa)
```

```{r}
#Loading data
load(here("data","datasets.Rdata"))
load(here("models_outputs","models_combined.Rdata")) # saved models results combined traits
load(here("models_outputs","models_traits.Rdata")) # saved models results separate traits
```

```{r}
#Loading auxiliary function for R2 calculations
source(here("scripts","r2_auxiliary_function.R"))
```

```{r}
# variance inflation factor function
vif.mer <- function (fit) {
    ## adapted from rms::vif
    v <- vcov(fit)
    nam <- names(fixef(fit))
    ## exclude intercepts
    ns <- sum(1 * (nam == "Intercept" | nam == "(Intercept)"))
    if (ns > 0) {
        v <- v[-(1:ns), -(1:ns), drop = FALSE]
        nam <- nam[-(1:ns)]
    }
    d <- diag(v)^0.5
    v <- diag(solve(v/(d %o% d)))
    names(v) <- nam
    v
}
```

# 1. Models with traits

**Specification of the models:** We used `lme4` package to perform a GLMM with binomial (proportion) distribution. An example of the code for each dataset are as follows:

```
mhigh.spe <- glmer(cbind(occor, n.visit-occor) ~  
                    forest_site400*lbody_size + 
                    forest_site400*nest +
                    forest_site400*diet + 
                    forest_site400*lower_stratum +
                    forest_land*lbody_size + 
                    forest_land*nest + 
                    forest_land*diet +
                    forest_land*lower_stratum + 
                    (forest_site400 + forest_land|sp) + 
                    (1|landscape:sp) + (1|site:sp) + 
                    (lbody_size + nest + diet + lower_stratum|landscape) +
                    (lbody_size + nest + diet + lower_stratum|site),
                    family=binomial, data=high.spe,
                    nAGQ = 1, control = glmerControl(optimizer = "bobyqa",
                                    optCtrl = list(maxfun = 500000)))
```

We ran separate models for each assemblage and trait. Afterwards, we ran one model with the combination of the traits body mass, diet, nest type and % of lower strata use. Table S3.\ref{tab:tabr2} shows the marginal R^2^ of all models terms.

```{r}
traits = c("lbody_size", "nest", "diet", "lower_stratum")
rhigh.spe <- rbind(r2.calc(mhigh.spe, atrib = traits) , 
                   r2.calc(bodyhigh.spe, atrib="lbody_size"), 
                   r2.calc(nesthigh.spe, atrib="nest"),
                   r2.calc(diethigh.spe, atrib="diet"),
                   r2.calc(frugihigh.spe, atrib="frugivory"),
                   r2.calc(insehigh.spe, atrib="insectivory"),
                   r2.calc(low_strathigh.spe, atrib = "lower_stratum"),
                   r2.calc(stratumhigh.spe, atrib="stratum"),
                   r2.calc(hwhigh.spe, atrib="handwing"))
rhigh.spe <- cbind(model = rep(c("Combined", "body mass", "nest type", 
                                 "main diet", "% frugivory", "% insetivory",
                                 "% lower strata", "foraging stratum",
                                 "hand-wing")),
                   rhigh.spe)  %>%
  mutate_if(is.numeric, function(x) 100*x)
colnames(rhigh.spe) <- c("Model", "Total", "trait*env",  "env|sp", "lands:sp",
                         "site:sp", "trait|lands", "trait|site")

rlow.spe <- rbind(r2.calc(mlow.spe, atrib = traits) , 
                   r2.calc(bodylow.spe, atrib="lbody_size"), 
                   r2.calc(nestlow.spe, atrib="nest"),
                   r2.calc(dietlow.spe, atrib="diet"),
                   r2.calc(frugilow.spe, atrib="frugivory"),
                   r2.calc(inselow.spe, atrib="insectivory"),
                   r2.calc(low_stratlow.spe, atrib = "lower_stratum"),
                   r2.calc(stratumlow.spe, atrib="stratum"),
                   r2.calc(hwlow.spe, atrib="handwing"))
rlow.spe <- cbind(model = rep(c("Combined", "body mass", "nest type", "main diet", 
                                 "% frugivory", "% insetivory","% lower strata", 
                                 "foraging stratum", "hand-wing")),
                   rlow.spe)  %>%
  mutate_if(is.numeric, function(x) 100*x)
colnames(rlow.spe) <- c("Model", "Total", "trait*env",  "env|sp", "lands:sp",
                         "site:sp", "trait|lands", "trait|site")
rhigh.gen <- rbind(r2.calc(mhigh.gen, atrib = traits) , 
                   r2.calc(bodyhigh.gen, atrib="lbody_size"), 
                   r2.calc(nesthigh.gen, atrib="nest"),
                   r2.calc(diethigh.gen, atrib="diet"),
                   r2.calc(frugihigh.gen, atrib="frugivory"),
                   r2.calc(insehigh.gen, atrib="insectivory"),
                   r2.calc(low_strathigh.gen, atrib = "lower_stratum"),
                   r2.calc(stratumhigh.gen, atrib="stratum"),
                   r2.calc(hwhigh.gen, atrib="handwing"))
rhigh.gen <- cbind(model = rep(c("Combined", "body mass", "nest type", "main diet", 
                                 "% frugivory", "% insetivory","% lower strata", 
                                 "foraging stratum", "hand-wing")),
                   rhigh.gen)  %>%
  mutate_if(is.numeric, function(x) 100*x)
colnames(rhigh.gen) <- c("Model", "Total", "trait*env",  "env|sp", "lands:sp",
                         "site:sp", "trait|lands", "trait|site")

rlow.gen <- rbind(r2.calc(mlow.gen, atrib = traits) , 
                  r2.calc(bodylow.gen, atrib="lbody_size"), 
                  r2.calc(nestlow.gen, atrib="nest"),
                  r2.calc(dietlow.gen, atrib="diet"),
                  r2.calc(frugilow.gen, atrib="frugivory"),
                  r2.calc(inselow.gen, atrib="insectivory"),
                  r2.calc(low_stratlow.gen, atrib = "lower_stratum"),
                  r2.calc(stratumlow.gen, atrib="stratum"),
                   r2.calc(hwlow.gen, atrib="handwing"))
rlow.gen <- cbind(model = rep(c("Combined", "body mass", "nest type", "main diet", 
                                "% frugivory", "% insetivory","% lower strata", 
                                "foraging stratum", "hand-wing")),
                  rlow.gen)  %>%
  mutate_if(is.numeric, function(x) 100*x)
colnames(rlow.gen) <- c("Model", "Total", "trait*env",  "env|sp", "lands:sp",
                        "site:sp", "trait|lands", "trait|site")
```

```{r, tabr2}
todos <- bind_rows(list(rhigh.spe, rlow.spe, rhigh.gen, rlow.gen))

kable(todos, "latex", booktabs=T, caption= "Overall and marginal R-squared of trait models in each dataset. For the marginal R-squared terms see Table 2 (main text).") %>%
   group_rows("Specialists", 1, 18) %>%
   group_rows("     High quality", 1,9) %>%
   group_rows("     Low quality", 10,18) %>%
   group_rows("Generalists", 19, 36) %>%
   group_rows("     High quality", 19,27) %>%
   group_rows("     Low quality", 28,36)
```

# 2. Models coeficients


Tables S3.\ref{tab:s3.2}, S3.\ref{tab:s3.3}, S3.\ref{tab:s3.4}, and S3.\ref{tab:s3.5} show the coefficients for each model.

```{r, s3.2}
 tidy(mhigh.spe, effects = "fixed") %>% 
  kable(digits=2,"latex", booktabs=T, caption = "Fixed effects coefficients for the model of specialists in high-quality matrix landscapes.")
```

```{r, s3.3}
 tidy(mlow.spe, effects = "fixed") %>% 
  kable(digits=2,"latex", booktabs=T,  caption = "Fixed effects coefficients for the model of specialists in low-quality matrix landscapes.")
```

```{r, s3.4}
 tidy(mhigh.gen, effects = "fixed") %>% 
  kable(digits=2, "latex", booktabs=T, caption = "Fixed effects coefficients for the model of generalists in high-quality matrix landscapes.")
```

```{r, s3.5}
 tidy(mlow.gen, effects = "fixed") %>% 
  kable(digits=2,"latex", booktabs=T, caption = "Fixed effects coefficients for the model of generalists in low-quality matrix landscapes.")
```

# 3. Models diagnostic

Variance Inflation Factor of the model parameters for each dataset in Table S3.\ref{tab:vif3}.

```{r vif3}
vifhigh.spe <- data.frame(termo = names(vif.mer(mhigh.spe2)),
                        high.spe =vif.mer(mhigh.spe2))
viflow.spe <- data.frame(termo = names(vif.mer(mlow.spe2)),
                        low.spe =vif.mer(mlow.spe2))
viflow.gen <- data.frame(termo = names(vif.mer(mhigh.gen2)),
                        low.gen =vif.mer(mhigh.gen2))
viflow.gen <- data.frame(termo = names(vif.mer(mlow.gen2)),
                        low.gen =vif.mer(mlow.gen2))
varif <- vifhigh.spe %>% full_join(viflow.spe, "termo") %>% 
  full_join(viflow.gen, "termo") %>% full_join(viflow.gen, "termo")
varif$termo <- c("forest.local", "body_mass", "nest_closed", "nest_open_semi",
                 "diet_insectivorous", "diet_onivorous", "lower_strata", 
                 "diet_granivorous", "forest.landscape", "diet_nectarivorous")
colnames(varif) <-c("parameter","High-quality", "Low-quality", "High-quality", "Low-quality")

kable(varif, "latex", booktabs=T, caption= "Variance Inflation Factor index for combined traits models in each dataset.", digits=2) %>%
 add_header_above(c(" " = 1, "Specialists" = 2, "Generalists"= 2))
```

\FloatBarrier

Example of the residual diagnostic of the model with the combined traits (main diet, body mass, nest type and % of lower strata use) for the forest specialists in high-quality matrix landscapes. The models' diagnostics for the other assemblages were all similar and can be checked in this Rmd file.


## Residual correlations among species and sites

Below we present the Kendall correlations for the residuals among species and sites for the models using the predictions for `site:sp` random effect (Observation Level Random Effect). For the residual correlations we followed the code provided by @miller_functional_2018.

```{r, fig.height=4}
nsites = 40
nspp = length(unique(high.spe$sp))
dat <- high.spe

dat$resid <- as.matrix(ranef(mhigh.spe)$`site:sp`)
X <- matrix(dat$resid, nrow=nsites, ncol=nspp, byrow=F)
rownames(X) <- unique(dat$ponto)
colnames(X) <- unique(dat$sp)

# species correlations
corrs.sp <- cor(X, method="kendall")
for(i in 1:nspp) corrs.sp[i,i] <- NA

# site correlations
corrs.site <- cor(t(X), method="kendall")
for(i in 1:nsites) corrs.site[i,i] <- NA
```

Range of species correlations: `r round(range(corrs.sp, na.rm=T),2)`. Range of sites correlations: `r round(range(corrs.site, na.rm=T),2)`. 

```{r, fig.height=4, fig.cap="Species residual Kendall correlations for the specialist species in high-quality matrix landscapes."}
corrplot(corrs.sp, method="square", type="upper",diag=F,tl.cex=0.5,cl.cex=0.5)
```

```{r, fig.height=3, fig.cap="Sites residual Kendall correlations for the specialist species in high-quality matrixlandscapes."}
corrplot(corrs.site, method="square", type="upper",diag=F,tl.cex=0.5,cl.cex=0.5)
```

```{r, fig.height=4, fig.cap= "Histograms of the residual Kendall correlations for the specialists species in high-quality matrix landscapes."}
x.sp <- matrix(corrs.sp, ncol=1)
x.site <- matrix(corrs.site, ncol=1)
x.sp <- x.sp[!is.na(x.sp)]
x.site <- x.site[!is.na(x.site)]

# Figure histogram
par(mfrow=c(1,2), mai=c(1,1,.5,.2))
hist(corrs.sp, main="", probability=T, breaks=.1*(-10:10), xlab="Correlation coefficients", ylab="Probability", ylim=c(0,4), cex.lab=1.2)
mtext("Species", side=3,cex=1.2)

hist(corrs.site, main="", probability=T, breaks=.1*(-10:10), xlab="Correlation coefficients", ylab="Probability", ylim=c(0,4), cex.lab=1.2)
mtext("Sites", side=3, cex=1.2)
```


<!--
### Forest specialist species in low-quality matrix landscapes

```{r, eval=F}
summary(mlow.spe)
```

```{r, fig.height=4}
nsites = 52
nspp = length(unique(low.spe$sp))
dat <- low.spe

dat$resid <- as.matrix(ranef(mlow.spe)$`site:sp`)
X <- matrix(dat$resid, nrow=nsites, ncol=nspp, byrow=F)
rownames(X) <- unique(dat$ponto)
colnames(X) <- unique(dat$sp)

# species correlations
corrs.sp <- cor(X, method="kendall")
for(i in 1:nspp) corrs.sp[i,i] <- NA

# site correlations
corrs.site <- cor(t(X), method="kendall")
for(i in 1:nsites) corrs.site[i,i] <- NA
```

Range of species correlations: `r round(range(corrs.sp, na.rm=T),2)`. Range of sites correlations: `r round(range(corrs.site, na.rm=T),2)`.

```{r, fig.height=4, fig.cap="Species residual Kendall correlations for the specialist species in low-quality matrix landscapes."}
corrplot(corrs.sp, method="square", type="upper",diag=F,tl.cex=0.5,cl.cex=0.5)
```

```{r, fig.height=3, fig.cap="Sites residual Kendall correlations for the specialist species in high-quality matrix landscapes."}
corrplot(corrs.site, method="square", type="upper",diag=F,tl.cex=0.5,cl.cex=0.5)
```

```{r, fig.height=4, fig.cap= "Histograms of the residual Kendall correlations for the specialists species in low-quality matrix landscapes."}
x.sp <- matrix(corrs.sp, ncol=1)
x.site <- matrix(corrs.site, ncol=1)
x.sp <- x.sp[!is.na(x.sp)]
x.site <- x.site[!is.na(x.site)]

# Figure histogram
par(mfrow=c(1,2), mai=c(1,1,.5,.2))
hist(corrs.sp, main="", probability=T, breaks=.1*(-10:10), xlab="Correlation coefficients", ylab="Probability", ylim=c(0,4), cex.lab=1.2)
mtext("Species", side=3,cex=1.2)

hist(corrs.site, main="", probability=T, breaks=.1*(-10:10), xlab="Correlation coefficients", ylab="Probability", ylim=c(0,4), cex.lab=1.2)
mtext("Sites", side=3, cex=1.2)
```

### Forest generalist species in high-quality matrix 

```{r}
summary(mhigh.gen)
```

```{r, fig.height=4}
nsites = 40
nspp = length(unique(high.gen$sp))
dat <- high.gen

dat$resid <- as.matrix(ranef(mhigh.gen)$`site:sp`)
X <- matrix(dat$resid, nrow=nsites, ncol=nspp, byrow=F)
rownames(X) <- unique(dat$ponto)
colnames(X) <- unique(dat$sp)

# species correlations
corrs.sp <- cor(X, method="kendall")
for(i in 1:nspp) corrs.sp[i,i] <- NA

# site correlations
corrs.site <- cor(t(X), method="kendall")
for(i in 1:nsites) corrs.site[i,i] <- NA
```

Range of species correlations: `r round(range(corrs.sp, na.rm=T),2)`. Range of sites correlations: `r round(range(corrs.site, na.rm=T),2)`.

```{r, fig.height=4, fig.cap="Species residual Kendall correlations for the generalist species in high-quality matrix landscapes."}
corrplot(corrs.sp, method="square", type="upper",diag=F,tl.cex=0.5,cl.cex=0.5)
```

```{r, fig.height=3, fig.cap="Sites residual Kendall correlations for the generalist species in high-quality matrix landscapes."}
corrplot(corrs.site, method="square", type="upper",diag=F,tl.cex=0.5,cl.cex=0.5)
```

```{r, fig.height=4, fig.cap= "Histograms of the residual Kendall correlations for the generalist species in high-quality matrix landscapes."}
x.sp <- matrix(corrs.sp, ncol=1)
x.site <- matrix(corrs.site, ncol=1)
x.sp <- x.sp[!is.na(x.sp)]
x.site <- x.site[!is.na(x.site)]

# Figure histogram
par(mfrow=c(1,2), mai=c(1,1,.5,.2))
hist(corrs.sp, main="", probability=T, breaks=.1*(-10:10), xlab="Correlation coefficients", ylab="Probability", ylim=c(0,4), cex.lab=1.2)
mtext("Species", side=3,cex=1.2)

hist(corrs.site, main="", probability=T, breaks=.1*(-10:10), xlab="Correlation coefficients", ylab="Probability", ylim=c(0,4), cex.lab=1.2)
mtext("Sites", side=3, cex=1.2)
```

### Forest generalist species in low-quality matrix landscapes

```{r}
summary(mlow.gen)
```

```{r, fig.height=4}
nsites = 52
nspp = length(unique(low.gen$sp))
dat <- low.gen

dat$resid <- as.matrix(ranef(mlow.gen)$`site:sp`)
X <- matrix(dat$resid, nrow=nsites, ncol=nspp, byrow=F)
rownames(X) <- unique(dat$ponto)
colnames(X) <- unique(dat$sp)

# species correlations
corrs.sp <- cor(X, method="kendall")
for(i in 1:nspp) corrs.sp[i,i] <- NA

# site correlations
corrs.site <- cor(t(X), method="kendall")
for(i in 1:nsites) corrs.site[i,i] <- NA
```

Range of species correlations: `r round(range(corrs.sp, na.rm=T),2)`. Range of sites correlations: `r round(range(corrs.site, na.rm=T),2)`. 

```{r, fig.height=4, fig.cap="Species residual Kendall correlations for the generalist species in low-quality matrix landscapes."}
corrplot(corrs.sp, method="square", type="upper",diag=F,tl.cex=0.5,cl.cex=0.5)
```

```{r, fig.height=3, fig.cap="Sites residual Kendall correlations for the generalist species in low-quality matrix landscapes."}
corrplot(corrs.site, method="square", type="upper",diag=F,tl.cex=0.5,cl.cex=0.5)
```

```{r, fig.height=4, fig.cap= "Histograms of the residual Kendall correlations for the generalists species in low-quality matrix landscapes."}
x.sp <- matrix(corrs.sp, ncol=1)
x.site <- matrix(corrs.site, ncol=1)
x.sp <- x.sp[!is.na(x.sp)]
x.site <- x.site[!is.na(x.site)]

# Figure histogram
par(mfrow=c(1,2), mai=c(1,1,.5,.2))
hist(corrs.sp, main="", probability=T, breaks=.1*(-10:10), xlab="Correlation coefficients", ylab="Probability", ylim=c(0,4), cex.lab=1.2)
mtext("Species", side=3,cex=1.2)

hist(corrs.site, main="", probability=T, breaks=.1*(-10:10), xlab="Correlation coefficients", ylab="Probability", ylim=c(0,4), cex.lab=1.2)
mtext("Sites", side=3, cex=1.2)
```

-->

## Residual diagnostic

We used DHARMa package (@hartig_dharma:_2018) for the diagnostic of quantile residuals. 

```{r, fig.cap= "Plots for model diagnostic form DHARMa package."}
resce <- simulateResiduals(mhigh.spe, n=1000, ref.form=~0)
plot(resce)
```

Residuals against predictors:

```{r, fig.height=8, fig.cap= "More for model diagnostic form DHARMa package."}
par(mfrow=c(4,2))
plotResiduals(resce, high.spe$forest_site400, xlab="Local forest cover 400m")
plotResiduals(resce, high.spe$forest_land, xlab="Landscape forest")
plotResiduals(resce, high.spe$lbody_size, xlab="Body size")
plotResiduals(resce, high.spe$nest, xlab="Nest type")
plotResiduals(resce, high.spe$diet, xlab="Main diet")
plotResiduals(resce, high.spe$low_strat, xlab="% lower strata use")
```


# Predictions for each species local forest cover

Landscape forest cover was fixed in 30%.

```{r}
#landscape cover 30%
fland = (30 - mean(high.spe$forest_landOrig))/
    sd(high.spe$forest_landOrig)
# local forest cover
f400st.high <- seq(min(high.spe$forest_site400), max(high.spe$forest_site400),
                     length.out=5)
f400Orig.high <- f400st.high * sd(high.spe$forest_site400Orig) +
                                              mean(high.spe$forest_site400Orig)
local.high <- data.frame(forest_site400=f400st.high, 
                         forest_site400Orig = f400Orig.high)

f400st.low <- seq(min(low.spe$forest_site400), max(low.spe$forest_site400),
                    length.out=5)
f400Orig.low <-f400st.low * sd(low.spe$forest_site400Orig) +
                  mean(low.spe$forest_site400Orig)
local.low <- data.frame(forest_site400=f400st.low, 
                         forest_site400Orig = f400Orig.low)
```


```{r}
# high quality specialists
sp.tra <- high.spe %>% select(sp, lbody_size, nest, diet, lower_stratum) %>%
  distinct() 
shigh.spe <- expand.grid(sp = unique(high.spe$sp),
                    forest_land = fland,
                    forest_site400 =f400st.high,
                    landscape = "P02",
                    site = "P02.P00") %>%
  left_join(sp.tra, "sp") %>%
  left_join(local.high, "forest_site400")
shigh.spe$pretes <- predict(mhigh.spe, newdata=shigh.spe, type="response",
                       re.form=NULL)

# low quality specialists
sp.tra2 <- low.spe %>% select(sp, lbody_size, nest, diet, lower_stratum) %>%
  distinct()
slow.spe <- expand.grid(sp = unique(low.spe$sp),
                    forest_land = fland,
                    forest_site400 =f400st.low,
                    #forest_site400Orig = f400Orig.low,
                    landscape = "148",
                    site = "148.P10") %>%
  left_join(sp.tra2, "sp")%>%
  left_join(local.low, "forest_site400")
slow.spe$pretes <- predict(mlow.spe, newdata=slow.spe, type="response",
                           re.form=NULL)
```

```{r, fig.height=18,fig.width=11, fig.cap="Forest specialist birds."}
bind_rows(list(High=shigh.spe,Low=slow.spe), .id="Matrix quality") %>%
  mutate(forest_site400Orig = round(forest_site400Orig)) %>%
  ggplot(aes(x=forest_site400Orig, y=pretes, col=`Matrix quality` )) +
  geom_line() +
  facet_wrap(~sp, ncol=8) + 
  theme_cowplot() +
  theme(legend.position = "top",
        strip.text.x = element_text(hjust=0, size=10))+
 scale_y_continuous(name= "Occurrence probability") +
  scale_x_continuous(name="Local forest cover (%)") +
  ggtitle("Forest specialists")
```

```{r}
# high quality generalists
sp.tra <- high.gen %>% select(sp, lbody_size, nest, diet, lower_stratum) %>%
  distinct() 
shigh.gen <- expand.grid(sp = unique(high.gen$sp),
                    forest_land = fland,
                    forest_site400 =f400st.high,
                    landscape = "P02",
                    site = "P02.P00") %>%
  left_join(sp.tra, "sp") %>%
  left_join(local.high, "forest_site400")
shigh.gen$pretes <- predict(mhigh.gen, newdata=shigh.gen, type="response",
                       re.form=NULL)

# low quality  generalists
sp.tra2 <- low.gen %>% select(sp, lbody_size, nest, diet, lower_stratum) %>%
  distinct()
slow.gen <- expand.grid(sp = unique(low.gen$sp),
                    forest_land = fland,
                    forest_site400 =f400st.low,
                    #forest_site400Orig = f400Orig.low,
                    landscape = "148",
                    site = "148.P10") %>%
  left_join(sp.tra2, "sp")%>%
  left_join(local.low, "forest_site400")
slow.gen$pretes <- predict(mlow.gen, newdata=slow.gen, type="response",
                           re.form=NULL)
```

```{r, fig.height=18,fig.width=11, fig.cap="Forest specialist birds."}
bind_rows(list(High=shigh.gen,Low=slow.gen), .id="Matrix quality") %>%
  mutate(forest_site400Orig = round(forest_site400Orig)) %>%
  ggplot(aes(x=forest_site400Orig, y=pretes, col=`Matrix quality` )) +
  geom_line() +
  facet_wrap(~sp, ncol=8) + 
  theme_cowplot() +
  theme(legend.position = "top",
        strip.text.x = element_text(hjust=0, size=10))+
 scale_y_continuous(name= "Occurrence probability") +
  scale_x_continuous(name="Local forest cover (%)") +
  ggtitle("Habitat generalists")
```



# References
